% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCSPHighWave.R
\name{addCSPHighWave}
\alias{addCSPHighWave}
\title{High-Wave Candlestick Pattern}
\usage{
addCSPHighWave(
  x,
  maxbodyCL = 0.1,
  minshadowCL = 0.3,
  output = c("xts", "tibble", "data.frame")
)
}
\arguments{
\item{x}{xts or data.frame with OHLC price data}

\item{maxbodyCL}{Maximum body ratio to candle length. Default 0.1.}

\item{minshadowCL}{Minimum shadow ratio to candle length. Default 0.3.}

\item{output}{Character. Return format: \code{"xts"} (default), \code{"tibble"}, or \code{"data.frame"}.}
}
\value{
A xts object containing the column(s):
\itemize{
\item HighWave: TRUE if pattern detected
}
}
\description{
Identifies High-Wave candles: small body with very long shadows on both sides.
}
\details{
Number of candle lines: \bold{1}
}
\seealso{
Other pattern-1bar:
\code{\link[=addCSPBeltHold]{addCSPBeltHold()}},
\code{\link[=addCSPClosingMarubozu]{addCSPClosingMarubozu()}},
\code{\link[=addCSPDoji]{addCSPDoji()}},
\code{\link[=addCSPHammer]{addCSPHammer()}},
\code{\link[=addCSPHangingMan]{addCSPHangingMan()}},
\code{\link[=addCSPInvertedHammer]{addCSPInvertedHammer()}},
\code{\link[=addCSPLongCandle]{addCSPLongCandle()}},
\code{\link[=addCSPLongCandleBody]{addCSPLongCandleBody()}},
\code{\link[=addCSPLongLeggedDoji]{addCSPLongLeggedDoji()}},
\code{\link[=addCSPMarubozu]{addCSPMarubozu()}},
\code{\link[=addCSPRickshawMan]{addCSPRickshawMan()}},
\code{\link[=addCSPShootingStar]{addCSPShootingStar()}},
\code{\link[=addCSPShortCandle]{addCSPShortCandle()}},
\code{\link[=addCSPShortCandleBody]{addCSPShortCandleBody()}},
\code{\link[=addCSPSpinningTop]{addCSPSpinningTop()}},
\code{\link[=addCSPTakuri]{addCSPTakuri()}},
\code{\link[=drawCSPBeltHold]{drawCSPBeltHold()}},
\code{\link[=drawCSPClosingMarubozu]{drawCSPClosingMarubozu()}},
\code{\link[=drawCSPDoji]{drawCSPDoji()}},
\code{\link[=drawCSPHammer]{drawCSPHammer()}},
\code{\link[=drawCSPHangingMan]{drawCSPHangingMan()}},
\code{\link[=drawCSPHighWave]{drawCSPHighWave()}},
\code{\link[=drawCSPInvertedHammer]{drawCSPInvertedHammer()}},
\code{\link[=drawCSPLongCandle]{drawCSPLongCandle()}},
\code{\link[=drawCSPLongCandleBody]{drawCSPLongCandleBody()}},
\code{\link[=drawCSPLongLeggedDoji]{drawCSPLongLeggedDoji()}},
\code{\link[=drawCSPMarubozu]{drawCSPMarubozu()}},
\code{\link[=drawCSPRickshawMan]{drawCSPRickshawMan()}},
\code{\link[=drawCSPShootingStar]{drawCSPShootingStar()}},
\code{\link[=drawCSPShortCandle]{drawCSPShortCandle()}},
\code{\link[=drawCSPShortCandleBody]{drawCSPShortCandleBody()}},
\code{\link[=drawCSPSpinningTop]{drawCSPSpinningTop()}},
\code{\link[=drawCSPTakuri]{drawCSPTakuri()}}
}
\concept{pattern-1bar}
