% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCSPNBlended.R
\name{addCSPNBlended}
\alias{addCSPNBlended}
\title{Calculate Open/High/Low/Close of N Blended Bars}
\usage{
addCSPNBlended(x, N, output = c("xts", "tibble", "data.frame"))
}
\arguments{
\item{x}{xts Time Series containing OHLC prices}

\item{N}{Number of bars to combine into one bar. Must be an integer >= 1.}

\item{output}{Character. Return format: \code{"xts"} (default), \code{"tibble"}, or \code{"data.frame"}.}
}
\value{
A xts object containing the columns:
\itemize{
\item {N}.Blended.Open: Opening price of the \code{N}-th elapsed candle
\item {N}.Blended.High: Highest high of the past \code{N} candles
\item {N}.Blended.Low: Lowest low of the past \code{N} candles
\item {N}.Blended.Close: Close price of the current candle
}
}
\description{
This function calculates OHLC prices of a combined candle representing N bars.
It creates a "blended" candle that aggregates price action over multiple periods.
}
\note{
This function is used by \code{\link{addPriceInfo}} to add price information
of detected candlestick patterns.
}
\examples{
\dontrun{
getSymbols("YHOO", adjust = TRUE)
addCSPNBlended(YHOO, N = 3) # combine 3 candles into one
}

}
\seealso{
\code{\link{addPriceInfo}}

Other pattern-nbar:
\code{\link[=addCSPNHigherClose]{addCSPNHigherClose()}},
\code{\link[=addCSPNLongBlackCandleBodies]{addCSPNLongBlackCandleBodies()}},
\code{\link[=addCSPNLongBlackCandles]{addCSPNLongBlackCandles()}},
\code{\link[=addCSPNLongWhiteCandleBodies]{addCSPNLongWhiteCandleBodies()}},
\code{\link[=addCSPNLongWhiteCandles]{addCSPNLongWhiteCandles()}},
\code{\link[=addCSPNLowerClose]{addCSPNLowerClose()}},
\code{\link[=drawCSPNBlended]{drawCSPNBlended()}},
\code{\link[=drawCSPNHigherClose]{drawCSPNHigherClose()}},
\code{\link[=drawCSPNLongBlackCandleBodies]{drawCSPNLongBlackCandleBodies()}},
\code{\link[=drawCSPNLongBlackCandles]{drawCSPNLongBlackCandles()}},
\code{\link[=drawCSPNLongWhiteCandleBodies]{drawCSPNLongWhiteCandleBodies()}},
\code{\link[=drawCSPNLongWhiteCandles]{drawCSPNLongWhiteCandles()}},
\code{\link[=drawCSPNLowerClose]{drawCSPNLowerClose()}}
}
\author{
Andreas Voellenklee
}
\concept{pattern-nbar}
