% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCSPSpinningTop.R
\name{addCSPSpinningTop}
\alias{addCSPSpinningTop}
\title{Spinning Top Candlestick Pattern}
\usage{
addCSPSpinningTop(
  x,
  maxbodyCL = 0.35,
  minbodyCL = 0.02,
  output = c("xts", "tibble", "data.frame")
)
}
\arguments{
\item{x}{xts or data.frame with OHLC price data}

\item{maxbodyCL}{Maximum body ratio. Default 0.35.}

\item{minbodyCL}{Minimum body ratio (not a doji). Default 0.02.}

\item{output}{Character. Return format: \code{"xts"} (default), \code{"tibble"}, or \code{"data.frame"}.}
}
\value{
A xts object containing the column(s):
\itemize{
\item BullSpinningTop: TRUE if pattern detected
\item BearSpinningTop: TRUE if pattern detected
}
}
\description{
Identifies Spinning Top candles: small body with shadows longer than the body.
}
\details{
Number of candle lines: \bold{1}
}
\seealso{
Other pattern-1bar:
\code{\link[=addCSPBeltHold]{addCSPBeltHold()}},
\code{\link[=addCSPClosingMarubozu]{addCSPClosingMarubozu()}},
\code{\link[=addCSPDoji]{addCSPDoji()}},
\code{\link[=addCSPHammer]{addCSPHammer()}},
\code{\link[=addCSPHangingMan]{addCSPHangingMan()}},
\code{\link[=addCSPHighWave]{addCSPHighWave()}},
\code{\link[=addCSPInvertedHammer]{addCSPInvertedHammer()}},
\code{\link[=addCSPLongCandle]{addCSPLongCandle()}},
\code{\link[=addCSPLongCandleBody]{addCSPLongCandleBody()}},
\code{\link[=addCSPLongLeggedDoji]{addCSPLongLeggedDoji()}},
\code{\link[=addCSPMarubozu]{addCSPMarubozu()}},
\code{\link[=addCSPRickshawMan]{addCSPRickshawMan()}},
\code{\link[=addCSPShootingStar]{addCSPShootingStar()}},
\code{\link[=addCSPShortCandle]{addCSPShortCandle()}},
\code{\link[=addCSPShortCandleBody]{addCSPShortCandleBody()}},
\code{\link[=addCSPTakuri]{addCSPTakuri()}},
\code{\link[=drawCSPBeltHold]{drawCSPBeltHold()}},
\code{\link[=drawCSPClosingMarubozu]{drawCSPClosingMarubozu()}},
\code{\link[=drawCSPDoji]{drawCSPDoji()}},
\code{\link[=drawCSPHammer]{drawCSPHammer()}},
\code{\link[=drawCSPHangingMan]{drawCSPHangingMan()}},
\code{\link[=drawCSPHighWave]{drawCSPHighWave()}},
\code{\link[=drawCSPInvertedHammer]{drawCSPInvertedHammer()}},
\code{\link[=drawCSPLongCandle]{drawCSPLongCandle()}},
\code{\link[=drawCSPLongCandleBody]{drawCSPLongCandleBody()}},
\code{\link[=drawCSPLongLeggedDoji]{drawCSPLongLeggedDoji()}},
\code{\link[=drawCSPMarubozu]{drawCSPMarubozu()}},
\code{\link[=drawCSPRickshawMan]{drawCSPRickshawMan()}},
\code{\link[=drawCSPShootingStar]{drawCSPShootingStar()}},
\code{\link[=drawCSPShortCandle]{drawCSPShortCandle()}},
\code{\link[=drawCSPShortCandleBody]{drawCSPShortCandleBody()}},
\code{\link[=drawCSPSpinningTop]{drawCSPSpinningTop()}},
\code{\link[=drawCSPTakuri]{drawCSPTakuri()}}

Other pattern-bull:
\code{\link[=addCSPAbandonedBaby]{addCSPAbandonedBaby()}},
\code{\link[=addCSPBeltHold]{addCSPBeltHold()}},
\code{\link[=addCSPBreakaway]{addCSPBreakaway()}},
\code{\link[=addCSPClosingMarubozu]{addCSPClosingMarubozu()}},
\code{\link[=addCSPConcealBabySwallow]{addCSPConcealBabySwallow()}},
\code{\link[=addCSPCounterAttack]{addCSPCounterAttack()}},
\code{\link[=addCSPDoji]{addCSPDoji()}},
\code{\link[=addCSPDojiStar]{addCSPDojiStar()}},
\code{\link[=addCSPEngulfing]{addCSPEngulfing()}},
\code{\link[=addCSPGap]{addCSPGap()}},
\code{\link[=addCSPGapSideSideWhite]{addCSPGapSideSideWhite()}},
\code{\link[=addCSPHammer]{addCSPHammer()}},
\code{\link[=addCSPHarami]{addCSPHarami()}},
\code{\link[=addCSPHomingPigeon]{addCSPHomingPigeon()}},
\code{\link[=addCSPInvertedHammer]{addCSPInvertedHammer()}},
\code{\link[=addCSPKicking]{addCSPKicking()}},
\code{\link[=addCSPLadderBottom]{addCSPLadderBottom()}},
\code{\link[=addCSPLongCandle]{addCSPLongCandle()}},
\code{\link[=addCSPLongCandleBody]{addCSPLongCandleBody()}},
\code{\link[=addCSPMarubozu]{addCSPMarubozu()}},
\code{\link[=addCSPMatchingLow]{addCSPMatchingLow()}},
\code{\link[=addCSPNHigherClose]{addCSPNHigherClose()}},
\code{\link[=addCSPNLongWhiteCandleBodies]{addCSPNLongWhiteCandleBodies()}},
\code{\link[=addCSPNLongWhiteCandles]{addCSPNLongWhiteCandles()}},
\code{\link[=addCSPPiercingPattern]{addCSPPiercingPattern()}},
\code{\link[=addCSPSeparatingLines]{addCSPSeparatingLines()}},
\code{\link[=addCSPShortCandle]{addCSPShortCandle()}},
\code{\link[=addCSPShortCandleBody]{addCSPShortCandleBody()}},
\code{\link[=addCSPStar]{addCSPStar()}},
\code{\link[=addCSPStickSandwich]{addCSPStickSandwich()}},
\code{\link[=addCSPStomach]{addCSPStomach()}},
\code{\link[=addCSPTakuri]{addCSPTakuri()}},
\code{\link[=addCSPTasukiGap]{addCSPTasukiGap()}},
\code{\link[=addCSPThreeInside]{addCSPThreeInside()}},
\code{\link[=addCSPThreeLineStrike]{addCSPThreeLineStrike()}},
\code{\link[=addCSPThreeMethods]{addCSPThreeMethods()}},
\code{\link[=addCSPThreeOutside]{addCSPThreeOutside()}},
\code{\link[=addCSPThreeStarsInSouth]{addCSPThreeStarsInSouth()}},
\code{\link[=addCSPThreeWhiteSoldiers]{addCSPThreeWhiteSoldiers()}},
\code{\link[=addCSPTristar]{addCSPTristar()}},
\code{\link[=addCSPUnique3River]{addCSPUnique3River()}},
\code{\link[=drawCSPAbandonedBaby]{drawCSPAbandonedBaby()}},
\code{\link[=drawCSPBeltHold]{drawCSPBeltHold()}},
\code{\link[=drawCSPBreakaway]{drawCSPBreakaway()}},
\code{\link[=drawCSPClosingMarubozu]{drawCSPClosingMarubozu()}},
\code{\link[=drawCSPConcealBabySwallow]{drawCSPConcealBabySwallow()}},
\code{\link[=drawCSPCounterAttack]{drawCSPCounterAttack()}},
\code{\link[=drawCSPDoji]{drawCSPDoji()}},
\code{\link[=drawCSPDojiStar]{drawCSPDojiStar()}},
\code{\link[=drawCSPEngulfing]{drawCSPEngulfing()}},
\code{\link[=drawCSPGap]{drawCSPGap()}},
\code{\link[=drawCSPGapSideSideWhite]{drawCSPGapSideSideWhite()}},
\code{\link[=drawCSPHammer]{drawCSPHammer()}},
\code{\link[=drawCSPHarami]{drawCSPHarami()}},
\code{\link[=drawCSPHomingPigeon]{drawCSPHomingPigeon()}},
\code{\link[=drawCSPInvertedHammer]{drawCSPInvertedHammer()}},
\code{\link[=drawCSPKicking]{drawCSPKicking()}},
\code{\link[=drawCSPLadderBottom]{drawCSPLadderBottom()}},
\code{\link[=drawCSPLongCandle]{drawCSPLongCandle()}},
\code{\link[=drawCSPLongCandleBody]{drawCSPLongCandleBody()}},
\code{\link[=drawCSPMarubozu]{drawCSPMarubozu()}},
\code{\link[=drawCSPMatchingLow]{drawCSPMatchingLow()}},
\code{\link[=drawCSPNHigherClose]{drawCSPNHigherClose()}},
\code{\link[=drawCSPNLongWhiteCandleBodies]{drawCSPNLongWhiteCandleBodies()}},
\code{\link[=drawCSPNLongWhiteCandles]{drawCSPNLongWhiteCandles()}},
\code{\link[=drawCSPPiercingPattern]{drawCSPPiercingPattern()}},
\code{\link[=drawCSPSeparatingLines]{drawCSPSeparatingLines()}},
\code{\link[=drawCSPShortCandle]{drawCSPShortCandle()}},
\code{\link[=drawCSPShortCandleBody]{drawCSPShortCandleBody()}},
\code{\link[=drawCSPSpinningTop]{drawCSPSpinningTop()}},
\code{\link[=drawCSPStar]{drawCSPStar()}},
\code{\link[=drawCSPStickSandwich]{drawCSPStickSandwich()}},
\code{\link[=drawCSPStomach]{drawCSPStomach()}},
\code{\link[=drawCSPTakuri]{drawCSPTakuri()}},
\code{\link[=drawCSPTasukiGap]{drawCSPTasukiGap()}},
\code{\link[=drawCSPThreeInside]{drawCSPThreeInside()}},
\code{\link[=drawCSPThreeLineStrike]{drawCSPThreeLineStrike()}},
\code{\link[=drawCSPThreeMethods]{drawCSPThreeMethods()}},
\code{\link[=drawCSPThreeOutside]{drawCSPThreeOutside()}},
\code{\link[=drawCSPThreeStarsInSouth]{drawCSPThreeStarsInSouth()}},
\code{\link[=drawCSPThreeWhiteSoldiers]{drawCSPThreeWhiteSoldiers()}},
\code{\link[=drawCSPTristar]{drawCSPTristar()}},
\code{\link[=drawCSPUnique3River]{drawCSPUnique3River()}}

Other pattern-bear:
\code{\link[=addCSPAbandonedBaby]{addCSPAbandonedBaby()}},
\code{\link[=addCSPAdvanceBlock]{addCSPAdvanceBlock()}},
\code{\link[=addCSPBeltHold]{addCSPBeltHold()}},
\code{\link[=addCSPBreakaway]{addCSPBreakaway()}},
\code{\link[=addCSPClosingMarubozu]{addCSPClosingMarubozu()}},
\code{\link[=addCSPCounterAttack]{addCSPCounterAttack()}},
\code{\link[=addCSPDarkCloudCover]{addCSPDarkCloudCover()}},
\code{\link[=addCSPDoji]{addCSPDoji()}},
\code{\link[=addCSPDojiStar]{addCSPDojiStar()}},
\code{\link[=addCSPEngulfing]{addCSPEngulfing()}},
\code{\link[=addCSPGap]{addCSPGap()}},
\code{\link[=addCSPGapSideSideWhite]{addCSPGapSideSideWhite()}},
\code{\link[=addCSPHangingMan]{addCSPHangingMan()}},
\code{\link[=addCSPHarami]{addCSPHarami()}},
\code{\link[=addCSPIdentical3Crows]{addCSPIdentical3Crows()}},
\code{\link[=addCSPInNeck]{addCSPInNeck()}},
\code{\link[=addCSPKicking]{addCSPKicking()}},
\code{\link[=addCSPLongCandle]{addCSPLongCandle()}},
\code{\link[=addCSPLongCandleBody]{addCSPLongCandleBody()}},
\code{\link[=addCSPMarubozu]{addCSPMarubozu()}},
\code{\link[=addCSPNLongBlackCandleBodies]{addCSPNLongBlackCandleBodies()}},
\code{\link[=addCSPNLongBlackCandles]{addCSPNLongBlackCandles()}},
\code{\link[=addCSPNLowerClose]{addCSPNLowerClose()}},
\code{\link[=addCSPOnNeck]{addCSPOnNeck()}},
\code{\link[=addCSPSeparatingLines]{addCSPSeparatingLines()}},
\code{\link[=addCSPShootingStar]{addCSPShootingStar()}},
\code{\link[=addCSPShortCandle]{addCSPShortCandle()}},
\code{\link[=addCSPShortCandleBody]{addCSPShortCandleBody()}},
\code{\link[=addCSPStalledPattern]{addCSPStalledPattern()}},
\code{\link[=addCSPStar]{addCSPStar()}},
\code{\link[=addCSPStomach]{addCSPStomach()}},
\code{\link[=addCSPTasukiGap]{addCSPTasukiGap()}},
\code{\link[=addCSPThreeBlackCrows]{addCSPThreeBlackCrows()}},
\code{\link[=addCSPThreeInside]{addCSPThreeInside()}},
\code{\link[=addCSPThreeLineStrike]{addCSPThreeLineStrike()}},
\code{\link[=addCSPThreeMethods]{addCSPThreeMethods()}},
\code{\link[=addCSPThreeOutside]{addCSPThreeOutside()}},
\code{\link[=addCSPThrusting]{addCSPThrusting()}},
\code{\link[=addCSPTristar]{addCSPTristar()}},
\code{\link[=addCSPTwoCrows]{addCSPTwoCrows()}},
\code{\link[=addCSPUpsideGap2Crows]{addCSPUpsideGap2Crows()}},
\code{\link[=drawCSPAbandonedBaby]{drawCSPAbandonedBaby()}},
\code{\link[=drawCSPAdvanceBlock]{drawCSPAdvanceBlock()}},
\code{\link[=drawCSPBeltHold]{drawCSPBeltHold()}},
\code{\link[=drawCSPBreakaway]{drawCSPBreakaway()}},
\code{\link[=drawCSPClosingMarubozu]{drawCSPClosingMarubozu()}},
\code{\link[=drawCSPCounterAttack]{drawCSPCounterAttack()}},
\code{\link[=drawCSPDarkCloudCover]{drawCSPDarkCloudCover()}},
\code{\link[=drawCSPDoji]{drawCSPDoji()}},
\code{\link[=drawCSPDojiStar]{drawCSPDojiStar()}},
\code{\link[=drawCSPEngulfing]{drawCSPEngulfing()}},
\code{\link[=drawCSPGap]{drawCSPGap()}},
\code{\link[=drawCSPGapSideSideWhite]{drawCSPGapSideSideWhite()}},
\code{\link[=drawCSPHangingMan]{drawCSPHangingMan()}},
\code{\link[=drawCSPHarami]{drawCSPHarami()}},
\code{\link[=drawCSPIdentical3Crows]{drawCSPIdentical3Crows()}},
\code{\link[=drawCSPInNeck]{drawCSPInNeck()}},
\code{\link[=drawCSPKicking]{drawCSPKicking()}},
\code{\link[=drawCSPLongCandle]{drawCSPLongCandle()}},
\code{\link[=drawCSPLongCandleBody]{drawCSPLongCandleBody()}},
\code{\link[=drawCSPMarubozu]{drawCSPMarubozu()}},
\code{\link[=drawCSPNLongBlackCandleBodies]{drawCSPNLongBlackCandleBodies()}},
\code{\link[=drawCSPNLongBlackCandles]{drawCSPNLongBlackCandles()}},
\code{\link[=drawCSPNLowerClose]{drawCSPNLowerClose()}},
\code{\link[=drawCSPOnNeck]{drawCSPOnNeck()}},
\code{\link[=drawCSPSeparatingLines]{drawCSPSeparatingLines()}},
\code{\link[=drawCSPShootingStar]{drawCSPShootingStar()}},
\code{\link[=drawCSPShortCandle]{drawCSPShortCandle()}},
\code{\link[=drawCSPShortCandleBody]{drawCSPShortCandleBody()}},
\code{\link[=drawCSPSpinningTop]{drawCSPSpinningTop()}},
\code{\link[=drawCSPStalledPattern]{drawCSPStalledPattern()}},
\code{\link[=drawCSPStar]{drawCSPStar()}},
\code{\link[=drawCSPStomach]{drawCSPStomach()}},
\code{\link[=drawCSPTasukiGap]{drawCSPTasukiGap()}},
\code{\link[=drawCSPThreeBlackCrows]{drawCSPThreeBlackCrows()}},
\code{\link[=drawCSPThreeInside]{drawCSPThreeInside()}},
\code{\link[=drawCSPThreeLineStrike]{drawCSPThreeLineStrike()}},
\code{\link[=drawCSPThreeMethods]{drawCSPThreeMethods()}},
\code{\link[=drawCSPThreeOutside]{drawCSPThreeOutside()}},
\code{\link[=drawCSPThrusting]{drawCSPThrusting()}},
\code{\link[=drawCSPTristar]{drawCSPTristar()}},
\code{\link[=drawCSPTwoCrows]{drawCSPTwoCrows()}},
\code{\link[=drawCSPUpsideGap2Crows]{drawCSPUpsideGap2Crows()}}
}
\concept{pattern-1bar}
\concept{pattern-bear}
\concept{pattern-bull}
