% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCSPStomach.R
\name{addCSPStomach}
\alias{addCSPStomach}
\title{Above/Below The Stomach Candlestick Pattern}
\usage{
addCSPStomach(x, output = c("xts", "tibble", "data.frame"))
}
\arguments{
\item{x}{xts Time Series containing Open and Close Prices}

\item{output}{Character. Return format: \code{"xts"} (default), \code{"tibble"}, or \code{"data.frame"}.}
}
\value{
A xts object containing the columns:
\itemize{
\item AboveTheStomach: TRUE if Above The Stomach pattern detected
\item BelowTheStomach: TRUE if Below The Stomach pattern detected
}
}
\description{
Identifies Above/Below The Stomach patterns in an Open/Close price series.
These are two-candle reversal patterns that signal potential trend changes.
}
\details{
Number of candle lines: \bold{2}

\strong{Above The Stomach:}
\itemize{
\item Prior trend: \bold{down}
\item A black candle followed by a white candle
\item The opening price of the second day must be at or above the midpoint of the first candle's body
\item This pattern signals a potential bullish reversal in a downtrend
}

\strong{Below The Stomach:}
\itemize{
\item Prior trend: \bold{up}
\item A white candle followed by a black candle
\item The opening price of the second day must be at or below the midpoint of the first candle's body
\item This pattern signals a potential bearish reversal in an uptrend
}
}
\note{
The function filters patterns that look like above/below the Stomach, without considering
the current trend direction. If only patterns in specific trends should be filtered,
an external trend detection function must be used. See examples.
}
\examples{
\dontrun{
getSymbols("YHOO", adjust = TRUE)
addCSPStomach(YHOO)

# Filter above the Stomach pattern in downtrends
addCSPStomach(YHOO)[, "AboveTheStomach"] &
  TrendDetectionChannel(lag(YHOO, k = 2))[, "DownTrend"]
}

}
\references{
The following sites were used to code/document this indicator:
\itemize{
\item \url{http://www.thepatternsite.com/AboveStomach.html}
\item \url{http://www.thepatternsite.com/BelowStomach.html}
}
}
\seealso{
Other pattern-2bar:
\code{\link[=addCSPDarkCloudCover]{addCSPDarkCloudCover()}},
\code{\link[=addCSPEngulfing]{addCSPEngulfing()}},
\code{\link[=addCSPGap]{addCSPGap()}},
\code{\link[=addCSPHarami]{addCSPHarami()}},
\code{\link[=addCSPInsideDay]{addCSPInsideDay()}},
\code{\link[=addCSPKicking]{addCSPKicking()}},
\code{\link[=addCSPOutsideDay]{addCSPOutsideDay()}},
\code{\link[=addCSPPiercingPattern]{addCSPPiercingPattern()}},
\code{\link[=addCSPTasukiGap]{addCSPTasukiGap()}},
\code{\link[=drawCSPDarkCloudCover]{drawCSPDarkCloudCover()}},
\code{\link[=drawCSPEngulfing]{drawCSPEngulfing()}},
\code{\link[=drawCSPGap]{drawCSPGap()}},
\code{\link[=drawCSPHarami]{drawCSPHarami()}},
\code{\link[=drawCSPInsideDay]{drawCSPInsideDay()}},
\code{\link[=drawCSPKicking]{drawCSPKicking()}},
\code{\link[=drawCSPOutsideDay]{drawCSPOutsideDay()}},
\code{\link[=drawCSPPiercingPattern]{drawCSPPiercingPattern()}},
\code{\link[=drawCSPStomach]{drawCSPStomach()}},
\code{\link[=drawCSPTasukiGap]{drawCSPTasukiGap()}}

Other pattern-bull:
\code{\link[=addCSPDoji]{addCSPDoji()}},
\code{\link[=addCSPEngulfing]{addCSPEngulfing()}},
\code{\link[=addCSPGap]{addCSPGap()}},
\code{\link[=addCSPHammer]{addCSPHammer()}},
\code{\link[=addCSPHarami]{addCSPHarami()}},
\code{\link[=addCSPInvertedHammer]{addCSPInvertedHammer()}},
\code{\link[=addCSPKicking]{addCSPKicking()}},
\code{\link[=addCSPLongCandle]{addCSPLongCandle()}},
\code{\link[=addCSPLongCandleBody]{addCSPLongCandleBody()}},
\code{\link[=addCSPMarubozu]{addCSPMarubozu()}},
\code{\link[=addCSPNHigherClose]{addCSPNHigherClose()}},
\code{\link[=addCSPNLongWhiteCandleBodies]{addCSPNLongWhiteCandleBodies()}},
\code{\link[=addCSPNLongWhiteCandles]{addCSPNLongWhiteCandles()}},
\code{\link[=addCSPPiercingPattern]{addCSPPiercingPattern()}},
\code{\link[=addCSPShortCandle]{addCSPShortCandle()}},
\code{\link[=addCSPShortCandleBody]{addCSPShortCandleBody()}},
\code{\link[=addCSPStar]{addCSPStar()}},
\code{\link[=addCSPTasukiGap]{addCSPTasukiGap()}},
\code{\link[=addCSPThreeInside]{addCSPThreeInside()}},
\code{\link[=addCSPThreeLineStrike]{addCSPThreeLineStrike()}},
\code{\link[=addCSPThreeMethods]{addCSPThreeMethods()}},
\code{\link[=addCSPThreeOutside]{addCSPThreeOutside()}},
\code{\link[=addCSPThreeWhiteSoldiers]{addCSPThreeWhiteSoldiers()}},
\code{\link[=drawCSPDoji]{drawCSPDoji()}},
\code{\link[=drawCSPEngulfing]{drawCSPEngulfing()}},
\code{\link[=drawCSPGap]{drawCSPGap()}},
\code{\link[=drawCSPHammer]{drawCSPHammer()}},
\code{\link[=drawCSPHarami]{drawCSPHarami()}},
\code{\link[=drawCSPInvertedHammer]{drawCSPInvertedHammer()}},
\code{\link[=drawCSPKicking]{drawCSPKicking()}},
\code{\link[=drawCSPLongCandle]{drawCSPLongCandle()}},
\code{\link[=drawCSPLongCandleBody]{drawCSPLongCandleBody()}},
\code{\link[=drawCSPMarubozu]{drawCSPMarubozu()}},
\code{\link[=drawCSPNHigherClose]{drawCSPNHigherClose()}},
\code{\link[=drawCSPNLongWhiteCandleBodies]{drawCSPNLongWhiteCandleBodies()}},
\code{\link[=drawCSPNLongWhiteCandles]{drawCSPNLongWhiteCandles()}},
\code{\link[=drawCSPPiercingPattern]{drawCSPPiercingPattern()}},
\code{\link[=drawCSPShortCandle]{drawCSPShortCandle()}},
\code{\link[=drawCSPShortCandleBody]{drawCSPShortCandleBody()}},
\code{\link[=drawCSPStar]{drawCSPStar()}},
\code{\link[=drawCSPStomach]{drawCSPStomach()}},
\code{\link[=drawCSPTasukiGap]{drawCSPTasukiGap()}},
\code{\link[=drawCSPThreeInside]{drawCSPThreeInside()}},
\code{\link[=drawCSPThreeLineStrike]{drawCSPThreeLineStrike()}},
\code{\link[=drawCSPThreeMethods]{drawCSPThreeMethods()}},
\code{\link[=drawCSPThreeOutside]{drawCSPThreeOutside()}},
\code{\link[=drawCSPThreeWhiteSoldiers]{drawCSPThreeWhiteSoldiers()}}

Other pattern-bear:
\code{\link[=addCSPDarkCloudCover]{addCSPDarkCloudCover()}},
\code{\link[=addCSPDoji]{addCSPDoji()}},
\code{\link[=addCSPEngulfing]{addCSPEngulfing()}},
\code{\link[=addCSPGap]{addCSPGap()}},
\code{\link[=addCSPHangingMan]{addCSPHangingMan()}},
\code{\link[=addCSPHarami]{addCSPHarami()}},
\code{\link[=addCSPKicking]{addCSPKicking()}},
\code{\link[=addCSPLongCandle]{addCSPLongCandle()}},
\code{\link[=addCSPLongCandleBody]{addCSPLongCandleBody()}},
\code{\link[=addCSPMarubozu]{addCSPMarubozu()}},
\code{\link[=addCSPNLongBlackCandleBodies]{addCSPNLongBlackCandleBodies()}},
\code{\link[=addCSPNLongBlackCandles]{addCSPNLongBlackCandles()}},
\code{\link[=addCSPNLowerClose]{addCSPNLowerClose()}},
\code{\link[=addCSPShootingStar]{addCSPShootingStar()}},
\code{\link[=addCSPShortCandle]{addCSPShortCandle()}},
\code{\link[=addCSPShortCandleBody]{addCSPShortCandleBody()}},
\code{\link[=addCSPStar]{addCSPStar()}},
\code{\link[=addCSPTasukiGap]{addCSPTasukiGap()}},
\code{\link[=addCSPThreeBlackCrows]{addCSPThreeBlackCrows()}},
\code{\link[=addCSPThreeInside]{addCSPThreeInside()}},
\code{\link[=addCSPThreeLineStrike]{addCSPThreeLineStrike()}},
\code{\link[=addCSPThreeMethods]{addCSPThreeMethods()}},
\code{\link[=addCSPThreeOutside]{addCSPThreeOutside()}},
\code{\link[=drawCSPDarkCloudCover]{drawCSPDarkCloudCover()}},
\code{\link[=drawCSPDoji]{drawCSPDoji()}},
\code{\link[=drawCSPEngulfing]{drawCSPEngulfing()}},
\code{\link[=drawCSPGap]{drawCSPGap()}},
\code{\link[=drawCSPHangingMan]{drawCSPHangingMan()}},
\code{\link[=drawCSPHarami]{drawCSPHarami()}},
\code{\link[=drawCSPKicking]{drawCSPKicking()}},
\code{\link[=drawCSPLongCandle]{drawCSPLongCandle()}},
\code{\link[=drawCSPLongCandleBody]{drawCSPLongCandleBody()}},
\code{\link[=drawCSPMarubozu]{drawCSPMarubozu()}},
\code{\link[=drawCSPNLongBlackCandleBodies]{drawCSPNLongBlackCandleBodies()}},
\code{\link[=drawCSPNLongBlackCandles]{drawCSPNLongBlackCandles()}},
\code{\link[=drawCSPNLowerClose]{drawCSPNLowerClose()}},
\code{\link[=drawCSPShootingStar]{drawCSPShootingStar()}},
\code{\link[=drawCSPShortCandle]{drawCSPShortCandle()}},
\code{\link[=drawCSPShortCandleBody]{drawCSPShortCandleBody()}},
\code{\link[=drawCSPStar]{drawCSPStar()}},
\code{\link[=drawCSPStomach]{drawCSPStomach()}},
\code{\link[=drawCSPTasukiGap]{drawCSPTasukiGap()}},
\code{\link[=drawCSPThreeBlackCrows]{drawCSPThreeBlackCrows()}},
\code{\link[=drawCSPThreeInside]{drawCSPThreeInside()}},
\code{\link[=drawCSPThreeLineStrike]{drawCSPThreeLineStrike()}},
\code{\link[=drawCSPThreeMethods]{drawCSPThreeMethods()}},
\code{\link[=drawCSPThreeOutside]{drawCSPThreeOutside()}}
}
\author{
Andreas Voellenklee
}
\concept{pattern-2bar}
\concept{pattern-bear}
\concept{pattern-bull}
