% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCSPTakuri.R
\name{addCSPTakuri}
\alias{addCSPTakuri}
\title{Takuri Candlestick Pattern}
\usage{
addCSPTakuri(
  x,
  maxbodyCL = 0.1,
  maxuppershadowCL = 0.1,
  minlowershadowCL = 0.6,
  output = c("xts", "tibble", "data.frame")
)
}
\arguments{
\item{x}{xts or data.frame with OHLC price data}

\item{maxbodyCL}{Maximum body ratio. Default 0.1.}

\item{maxuppershadowCL}{Maximum upper shadow ratio. Default 0.1.}

\item{minlowershadowCL}{Minimum lower shadow ratio. Default 0.6.}

\item{output}{Character. Return format: \code{"xts"} (default), \code{"tibble"}, or \code{"data.frame"}.}
}
\value{
A xts object containing the column(s):
\itemize{
\item Takuri: TRUE if pattern detected
}
}
\description{
Identifies Takuri (water-divining stick): dragonfly doji with very long lower shadow.
}
\details{
Number of candle lines: \bold{1}
}
\seealso{
Other pattern-1bar:
\code{\link[=addCSPBeltHold]{addCSPBeltHold()}},
\code{\link[=addCSPClosingMarubozu]{addCSPClosingMarubozu()}},
\code{\link[=addCSPDoji]{addCSPDoji()}},
\code{\link[=addCSPHammer]{addCSPHammer()}},
\code{\link[=addCSPHangingMan]{addCSPHangingMan()}},
\code{\link[=addCSPHighWave]{addCSPHighWave()}},
\code{\link[=addCSPInvertedHammer]{addCSPInvertedHammer()}},
\code{\link[=addCSPLongCandle]{addCSPLongCandle()}},
\code{\link[=addCSPLongCandleBody]{addCSPLongCandleBody()}},
\code{\link[=addCSPLongLeggedDoji]{addCSPLongLeggedDoji()}},
\code{\link[=addCSPMarubozu]{addCSPMarubozu()}},
\code{\link[=addCSPRickshawMan]{addCSPRickshawMan()}},
\code{\link[=addCSPShootingStar]{addCSPShootingStar()}},
\code{\link[=addCSPShortCandle]{addCSPShortCandle()}},
\code{\link[=addCSPShortCandleBody]{addCSPShortCandleBody()}},
\code{\link[=addCSPSpinningTop]{addCSPSpinningTop()}},
\code{\link[=drawCSPBeltHold]{drawCSPBeltHold()}},
\code{\link[=drawCSPClosingMarubozu]{drawCSPClosingMarubozu()}},
\code{\link[=drawCSPDoji]{drawCSPDoji()}},
\code{\link[=drawCSPHammer]{drawCSPHammer()}},
\code{\link[=drawCSPHangingMan]{drawCSPHangingMan()}},
\code{\link[=drawCSPHighWave]{drawCSPHighWave()}},
\code{\link[=drawCSPInvertedHammer]{drawCSPInvertedHammer()}},
\code{\link[=drawCSPLongCandle]{drawCSPLongCandle()}},
\code{\link[=drawCSPLongCandleBody]{drawCSPLongCandleBody()}},
\code{\link[=drawCSPLongLeggedDoji]{drawCSPLongLeggedDoji()}},
\code{\link[=drawCSPMarubozu]{drawCSPMarubozu()}},
\code{\link[=drawCSPRickshawMan]{drawCSPRickshawMan()}},
\code{\link[=drawCSPShootingStar]{drawCSPShootingStar()}},
\code{\link[=drawCSPShortCandle]{drawCSPShortCandle()}},
\code{\link[=drawCSPShortCandleBody]{drawCSPShortCandleBody()}},
\code{\link[=drawCSPSpinningTop]{drawCSPSpinningTop()}},
\code{\link[=drawCSPTakuri]{drawCSPTakuri()}}

Other pattern-bull:
\code{\link[=addCSPAbandonedBaby]{addCSPAbandonedBaby()}},
\code{\link[=addCSPBeltHold]{addCSPBeltHold()}},
\code{\link[=addCSPBreakaway]{addCSPBreakaway()}},
\code{\link[=addCSPClosingMarubozu]{addCSPClosingMarubozu()}},
\code{\link[=addCSPConcealBabySwallow]{addCSPConcealBabySwallow()}},
\code{\link[=addCSPCounterAttack]{addCSPCounterAttack()}},
\code{\link[=addCSPDoji]{addCSPDoji()}},
\code{\link[=addCSPDojiStar]{addCSPDojiStar()}},
\code{\link[=addCSPEngulfing]{addCSPEngulfing()}},
\code{\link[=addCSPGap]{addCSPGap()}},
\code{\link[=addCSPGapSideSideWhite]{addCSPGapSideSideWhite()}},
\code{\link[=addCSPHammer]{addCSPHammer()}},
\code{\link[=addCSPHarami]{addCSPHarami()}},
\code{\link[=addCSPHomingPigeon]{addCSPHomingPigeon()}},
\code{\link[=addCSPInvertedHammer]{addCSPInvertedHammer()}},
\code{\link[=addCSPKicking]{addCSPKicking()}},
\code{\link[=addCSPLadderBottom]{addCSPLadderBottom()}},
\code{\link[=addCSPLongCandle]{addCSPLongCandle()}},
\code{\link[=addCSPLongCandleBody]{addCSPLongCandleBody()}},
\code{\link[=addCSPMarubozu]{addCSPMarubozu()}},
\code{\link[=addCSPMatchingLow]{addCSPMatchingLow()}},
\code{\link[=addCSPNHigherClose]{addCSPNHigherClose()}},
\code{\link[=addCSPNLongWhiteCandleBodies]{addCSPNLongWhiteCandleBodies()}},
\code{\link[=addCSPNLongWhiteCandles]{addCSPNLongWhiteCandles()}},
\code{\link[=addCSPPiercingPattern]{addCSPPiercingPattern()}},
\code{\link[=addCSPSeparatingLines]{addCSPSeparatingLines()}},
\code{\link[=addCSPShortCandle]{addCSPShortCandle()}},
\code{\link[=addCSPShortCandleBody]{addCSPShortCandleBody()}},
\code{\link[=addCSPSpinningTop]{addCSPSpinningTop()}},
\code{\link[=addCSPStar]{addCSPStar()}},
\code{\link[=addCSPStickSandwich]{addCSPStickSandwich()}},
\code{\link[=addCSPStomach]{addCSPStomach()}},
\code{\link[=addCSPTasukiGap]{addCSPTasukiGap()}},
\code{\link[=addCSPThreeInside]{addCSPThreeInside()}},
\code{\link[=addCSPThreeLineStrike]{addCSPThreeLineStrike()}},
\code{\link[=addCSPThreeMethods]{addCSPThreeMethods()}},
\code{\link[=addCSPThreeOutside]{addCSPThreeOutside()}},
\code{\link[=addCSPThreeStarsInSouth]{addCSPThreeStarsInSouth()}},
\code{\link[=addCSPThreeWhiteSoldiers]{addCSPThreeWhiteSoldiers()}},
\code{\link[=addCSPTristar]{addCSPTristar()}},
\code{\link[=addCSPUnique3River]{addCSPUnique3River()}},
\code{\link[=drawCSPAbandonedBaby]{drawCSPAbandonedBaby()}},
\code{\link[=drawCSPBeltHold]{drawCSPBeltHold()}},
\code{\link[=drawCSPBreakaway]{drawCSPBreakaway()}},
\code{\link[=drawCSPClosingMarubozu]{drawCSPClosingMarubozu()}},
\code{\link[=drawCSPConcealBabySwallow]{drawCSPConcealBabySwallow()}},
\code{\link[=drawCSPCounterAttack]{drawCSPCounterAttack()}},
\code{\link[=drawCSPDoji]{drawCSPDoji()}},
\code{\link[=drawCSPDojiStar]{drawCSPDojiStar()}},
\code{\link[=drawCSPEngulfing]{drawCSPEngulfing()}},
\code{\link[=drawCSPGap]{drawCSPGap()}},
\code{\link[=drawCSPGapSideSideWhite]{drawCSPGapSideSideWhite()}},
\code{\link[=drawCSPHammer]{drawCSPHammer()}},
\code{\link[=drawCSPHarami]{drawCSPHarami()}},
\code{\link[=drawCSPHomingPigeon]{drawCSPHomingPigeon()}},
\code{\link[=drawCSPInvertedHammer]{drawCSPInvertedHammer()}},
\code{\link[=drawCSPKicking]{drawCSPKicking()}},
\code{\link[=drawCSPLadderBottom]{drawCSPLadderBottom()}},
\code{\link[=drawCSPLongCandle]{drawCSPLongCandle()}},
\code{\link[=drawCSPLongCandleBody]{drawCSPLongCandleBody()}},
\code{\link[=drawCSPMarubozu]{drawCSPMarubozu()}},
\code{\link[=drawCSPMatchingLow]{drawCSPMatchingLow()}},
\code{\link[=drawCSPNHigherClose]{drawCSPNHigherClose()}},
\code{\link[=drawCSPNLongWhiteCandleBodies]{drawCSPNLongWhiteCandleBodies()}},
\code{\link[=drawCSPNLongWhiteCandles]{drawCSPNLongWhiteCandles()}},
\code{\link[=drawCSPPiercingPattern]{drawCSPPiercingPattern()}},
\code{\link[=drawCSPSeparatingLines]{drawCSPSeparatingLines()}},
\code{\link[=drawCSPShortCandle]{drawCSPShortCandle()}},
\code{\link[=drawCSPShortCandleBody]{drawCSPShortCandleBody()}},
\code{\link[=drawCSPSpinningTop]{drawCSPSpinningTop()}},
\code{\link[=drawCSPStar]{drawCSPStar()}},
\code{\link[=drawCSPStickSandwich]{drawCSPStickSandwich()}},
\code{\link[=drawCSPStomach]{drawCSPStomach()}},
\code{\link[=drawCSPTakuri]{drawCSPTakuri()}},
\code{\link[=drawCSPTasukiGap]{drawCSPTasukiGap()}},
\code{\link[=drawCSPThreeInside]{drawCSPThreeInside()}},
\code{\link[=drawCSPThreeLineStrike]{drawCSPThreeLineStrike()}},
\code{\link[=drawCSPThreeMethods]{drawCSPThreeMethods()}},
\code{\link[=drawCSPThreeOutside]{drawCSPThreeOutside()}},
\code{\link[=drawCSPThreeStarsInSouth]{drawCSPThreeStarsInSouth()}},
\code{\link[=drawCSPThreeWhiteSoldiers]{drawCSPThreeWhiteSoldiers()}},
\code{\link[=drawCSPTristar]{drawCSPTristar()}},
\code{\link[=drawCSPUnique3River]{drawCSPUnique3River()}}
}
\concept{pattern-1bar}
\concept{pattern-bull}
