% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPriceInfo.R
\name{addPriceInfo}
\alias{addPriceInfo}
\title{Add OHLC Information for Detected Candlestick Patterns}
\usage{
addPriceInfo(TS, CSP)
}
\arguments{
\item{TS}{An xts time series containing OHLC prices}

\item{CSP}{An xts time series of detected candlestick patterns, based on TS}
}
\value{
An xts object containing the columns:
\itemize{
\item All columns of CSP
\item \code{Formation.Open}: Opening price of the detected candlestick formation
\item \code{Formation.High}: Highest high of the detected candlestick formation
\item \code{Formation.Low}: Lowest low of the detected candlestick formation
\item \code{Formation.Close}: Close price of the detected candlestick formation
}
}
\description{
This function returns not only the occurrences of candlestick patterns, but also the OHLC price information
of the pattern's formation. The OHLC price information is only set on dates when the pattern is detected,
and 0 otherwise.
}
\details{
The xtsAttribute \code{bars} of \code{CSP} is used to calculate the blended candlestick of the formation.
}
\examples{
\dontrun{
getSymbols("YHOO", adjust = TRUE)

# Return detected Engulfing Pattern
CSPEngulfing(YHOO)

# Return detected Engulfing Pattern including formation's OHLC
addPriceInfo(YHOO, CSPEngulfing(YHOO))
}

}
\seealso{
\code{\link{CSPNBlended}}
}
\author{
Andreas Voellenklee
}
