% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSPDarkCloudCover.R
\name{CSPDarkCloudCover}
\alias{CSPDarkCloudCover}
\title{Dark Cloud Cover Candlestick Pattern}
\usage{
CSPDarkCloudCover(TS, n = 20, minbodysizeMedian = 1)
}
\arguments{
\item{TS}{xts Time Series containing Open and Close prices}

\item{n}{Number of preceding candles to calculate median candle body length. Default is 20.}

\item{minbodysizeMedian}{Minimum body length relative to the median of the past \code{n} body sizes. Default is 1.}
}
\value{
A xts object containing the column:
\itemize{
\item DarkCloudCover: TRUE if Dark Cloud Cover pattern detected
}
}
\description{
Identifies Dark Cloud Cover patterns in an Open/Close price series.
This is a two-candle bearish reversal pattern that occurs during an uptrend.
}
\details{
Number of candle lines: \bold{2}

A Dark Cloud Cover pattern consists of:
\enumerate{
\item A long white candlestick (body length > median of past n candles)
\item A gap higher opening on the next day
\item A black candlestick that closes more than halfway into the prior white candlestick's body
\item The close of the second candle is still above the open of the first candle
}

This pattern is considered a bearish reversal signal when it appears during an uptrend.
}
\note{
The function filters patterns that look like dark cloud covers, without considering
the current trend direction. If only patterns in uptrends should be filtered,
an external trend detection function must be used. See examples.
}
\examples{
\dontrun{
getSymbols("YHOO", adjust = TRUE)
CSPDarkCloudCover(YHOO)

# Filter dark cloud covers that occur in uptrends.
# The lag of 2 periods of the time series for trend detection
# ensures that the uptrend is active before the
# dark cloud cover occurs.
CSPDarkCloudCover(YHOO) &
  TrendDetectionChannel(lag(YHOO, k = 2))[, "UpTrend"]
}

}
\references{
The following sites were used to code/document this indicator:
\itemize{
\item \url{http://www.candlesticker.com/Bearish.asp}
\item \url{http://www.onlinetradingconcepts.com/TechnicalAnalysis/Candlesticks/DarkCloudCover.html}
}
}
\seealso{
The counterpart of this pattern is \code{\link{CSPPiercingPattern}}
}
\author{
Andreas Voellenklee
}
