% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSPDoji.R
\name{CSPDoji}
\alias{CSPDoji}
\title{Doji Candlestick Pattern}
\usage{
CSPDoji(TS, maxbodyCL = 0.1, maxshadowCL = 0.1)
}
\arguments{
\item{TS}{An xts time series containing Open and Close prices}

\item{maxbodyCL}{Maximum body-to-length ratio, default = 0.1}

\item{maxshadowCL}{Maximum tolerated upper (lower) shadow ratio to identify a dragonfly (gravestone) doji, default = 0.1}
}
\value{
An xts object containing the following columns:
\itemize{
\item \code{Doji}: TRUE if a Doji pattern is detected
\item \code{DragonflyDoji}: TRUE if a Dragonfly Doji pattern is detected
\item \code{GravestoneDoji}: TRUE if a Gravestone Doji pattern is detected
}
}
\description{
Identify Doji Patterns in an OHLC price series
}
\details{
Number of candle lines: \strong{1}

A Doji is formed when the open and close prices are the same or very close. By default settings, a Doji is identified when the body of the candle is smaller than or equal to 1/10 of its full length.

The bullish Dragonfly Doji pattern has a long lower shadow but almost no upper shadow. Conversely, the bearish Gravestone Doji has a long upper shadow but almost no lower shadow. Dragonfly Doji and Gravestone Doji can be reversal signals during downtrends/uptrends.
}
\note{
This function only filters candlesticks that resemble Doji patterns, without considering the current trend direction. To filter for Doji patterns in an uptrend, an external trend detection function must be used. See examples.
}
\examples{
\dontrun{
# Obtain example data
getSymbols("YHOO", adjust = TRUE)

# Identify Doji patterns
CSPDoji(YHOO)

# Filter for Doji patterns where the open price equals the close price
CSPDoji(YHOO, maxbodyCL = 0)

# Filter for Gravestone Doji patterns in an uptrend
CSPDoji(YHOO)[, "GravestoneDoji"] & TrendDetectionChannel(YHOO)[, "UpTrend"]
}

}
\references{
Reference sites used for coding/documenting this candlestick pattern:
\itemize{
\item \url{http://www.onlinetradingconcepts.com/TechnicalAnalysis/Candlesticks/Doji.html}
\item \url{http://www.candlesticker.com/Bullish.asp}
\item \url{http://www.candlesticker.com/Bearish.asp}
}
}
\author{
Andreas Voellenklee
}
