% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSPEngulfing.R
\name{CSPEngulfing}
\alias{CSPEngulfing}
\title{Engulfing Candlestick Pattern}
\usage{
CSPEngulfing(TS)
}
\arguments{
\item{TS}{xts Time Series containing Open and Close prices}
}
\value{
A xts object containing the columns:
\itemize{
\item Bull.Engulfing: TRUE if bullish engulfing pattern detected
\item Bear.Engulfing: TRUE if bearish engulfing pattern detected
}
}
\description{
Identifies Bullish and Bearish Engulfing Patterns in an Open/Close price series.
}
\details{
Number of candle lines: \bold{2}

\strong{Bullish Engulfing Pattern:}
\itemize{
\item Prior trend: \bold{down}
\item Characterized by a large white real body engulfing a preceding small black real body
\item The white body does not necessarily engulf the shadows of the black body but totally engulfs the body itself
\item Appears during a downtrend
}

\strong{Bearish Engulfing Pattern:}
\itemize{
\item Prior trend: \bold{up}
\item Characterized by a large black real body engulfing a preceding small white real body
\item The black body does not necessarily engulf the shadows of the white body but totally engulfs the body itself
\item Appears during an uptrend
}
}
\note{
The function filters patterns that look like engulfing patterns, without considering
the current trend direction. If only patterns in uptrends/downtrends should be filtered,
an external trend detection function must be used. See examples.
}
\examples{
\dontrun{
getSymbols("YHOO", adjust = TRUE)
CSPEngulfing(YHOO)

# Filter bearish engulfing patterns that occur in uptrends
CSPEngulfing(YHOO)[, "Bear.Engulfing"] &
  TrendDetectionChannel(YHOO)[, "UpTrend"]
}

}
\references{
The following sites were used to code/document this candlestick pattern:
\itemize{
\item \url{http://thepatternsite.com/BullEngulfing.html}
\item \url{http://www.onlinetradingconcepts.com/TechnicalAnalysis/Candlesticks/BullishEngulfing.html}
\item \url{http://www.candlesticker.com/Bullish.asp}
\item \url{http://www.candlesticker.com/Bearish.asp}
}
}
\author{
Andreas Voellenklee
}
