% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSPHammer.R
\name{CSPHammer}
\alias{CSPHammer}
\title{Hammer Candlestick Pattern}
\usage{
CSPHammer(TS, minlowershadowCL = 2/3, maxuppershadowCL = 0.1, minbodyCL = 0.1)
}
\arguments{
\item{TS}{xts Time Series containing Open, High, Low and Close Prices}

\item{minlowershadowCL}{Minimum lower shadow to candle length ratio. Default is 2/3.}

\item{maxuppershadowCL}{Maximum tolerated upper shadow to candle length ratio. Default is 0.1.}

\item{minbodyCL}{Minimum body to candle length ratio. Default is 0.1.}
}
\value{
A xts object containing the column:
\itemize{
\item Hammer: TRUE if hammer pattern detected
}
}
\description{
Identifies Hammer patterns in an OHLC price series.
The hammer is a one-day bullish reversal pattern that appears during a downtrend.
}
\details{
Number of candle lines: \bold{1}

The hammer is a one-day formation expected to be an early sign for the reversal
of a downtrend into an uptrend. It has:
\itemize{
\item A long lower shadow (at least \code{minlowershadowCL} of the candle length)
\item A small body at the top of the candle (at least \code{minbodyCL} of the candle length)
\item No or only a very short upper shadow (at most \code{maxuppershadowCL} of the candle length)
\item The color of the body is not important
}
}
\note{
The function filters candles that look like hammers, without considering
the current trend direction. If only hammer patterns in a downtrend should be
filtered, an external trend detection function must be used. See examples.
}
\examples{
\dontrun{
getSymbols("YHOO", adjust = TRUE)

# Filter for hammer patterns
CSPHammer(YHOO)

# Filter for hammer patterns that occur in downtrends
CSPHammer(YHOO) & TrendDetectionChannel(YHOO)[, "DownTrend"]
}

}
\references{
The following site was used to code/document this candlestick pattern:
\url{http://www.candlesticker.com/Cs18.asp}
}
\seealso{
\code{\link{CSPDoji}}
\code{\link{TrendDetectionChannel}}
\code{\link{TrendDetectionSMA}}
}
\author{
Andreas Voellenklee
}
