% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSPKicking.R
\name{CSPKicking}
\alias{CSPKicking}
\title{Kicking Candlestick Pattern}
\usage{
CSPKicking(
  TS,
  ignoreShadows = TRUE,
  n = 20,
  threshold = 1,
  ATRFactor = 1,
  maxshadowCL = 0.1
)
}
\arguments{
\item{TS}{xts Time Series containing OHLC prices}

\item{ignoreShadows}{If TRUE, only Open and Close Price are evaluated.
High and Low are ignored. Default is TRUE.}

\item{n}{Number of preceding candles to calculate median candle length or average True Range. Default is 20.}

\item{threshold}{Minimum candle length in relation to the median candle length of \code{n} preceding candles. Default is 1.}

\item{ATRFactor}{Minimum size of candle body compared to the ATR. Default is 1.}

\item{maxshadowCL}{Maximum tolerated upper and lower shadow to candle length ratio. Default is 0.1.}
}
\value{
A xts object containing the columns:
\itemize{
\item Bull.Kicking: TRUE if Bullish Kicking pattern detected
\item Bear.Kicking: TRUE if Bearish Kicking pattern detected
}
}
\description{
Identifies bullish and bearish Kicking candlestick patterns in an OHLC price series.
This is a two-candle pattern that signals strong reversal potential.
}
\details{
Number of candle lines: \bold{2}

\strong{Bullish Kicking Pattern:}
\itemize{
\item A White Marubozu following a Black Marubozu
\item After the Black Marubozu, market gaps higher on the opening
\item The pattern is filtered with \code{ignoreShadows=FALSE} and occurs very scarcely
\item When \code{ignoreShadows=TRUE}, a less strict filter is applied: a long black candle body
followed by a long white candle body with a gap
}

\strong{Bearish Kicking Pattern:}
\itemize{
\item Opposite of Bullish Kicking Pattern
\item A Black Marubozu following a White Marubozu with a gap down
}

Unlike most other candlestick patterns, the previous market direction is not important for this pattern.
}
\examples{
\dontrun{
getSymbols("YHOO", adjust = TRUE)

# Look for Kicking Pattern right out of the textbook
# They occur only once in a blue moon
CSPKicking(YHOO, ignoreShadows = FALSE, maxshadowCL = 0)

# Use less strict filter rules
CSPKicking(YHOO)
}

}
\references{
The following sites were used to code/document this indicator:
\itemize{
\item \url{http://www.candlesticker.com/Bullish.asp}
\item \url{http://www.candlesticker.com/Bearish.asp}
}
}
\seealso{
\code{\link{CSPLongCandleBody}}
\code{\link{CSPMarubozu}}
\code{\link{CSPGap}}
}
\author{
Andreas Voellenklee
}
