% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSPLongCandle.R
\name{CSPLongCandle}
\alias{CSPLongCandle}
\title{Long Candlestick Pattern}
\usage{
CSPLongCandle(TS, n = 20, threshold = 1)
}
\arguments{
\item{TS}{xts Time Series containing OHLC prices}

\item{n}{Number of preceding candles to calculate median candle length. Default is 20.}

\item{threshold}{Minimum candle length in relation to the median candle length
of \code{n} preceding candles. Default is 1.}
}
\value{
A xts object containing the columns:
\itemize{
\item LongWhiteCandle: TRUE if Long White Candle detected
\item LongBlackCandle: TRUE if Long Black Candle detected
}
}
\description{
Identifies long candlesticks in an OHLC price series based on their relative length
compared to the median of preceding candles.
}
\details{
Classifies candles as long based on their relative length compared to the median
of the relative candle lengths of the preceding \code{n} candles.

Relative candle length is calculated as \code{(Hi-Lo)/((Hi+Lo)/2)}.
}
\examples{
\dontrun{
getSymbols("YHOO", adjust = TRUE)
CSPLongCandle(YHOO)
}

}
\seealso{
\code{\link{CSPNLongWhiteCandles}}
}
\author{
Andreas Voellenklee
}
