% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSPMarubozu.R
\name{CSPMarubozu}
\alias{CSPMarubozu}
\title{Marubozu Candlestick Pattern}
\usage{
CSPMarubozu(
  TS,
  n = 20,
  ATRFactor = 1,
  maxuppershadowCL = 0.1,
  maxlowershadowCL = 0.1
)
}
\arguments{
\item{TS}{xts Time Series containing OHLC prices}

\item{n}{Number of preceding candles to calculate Average True Range. Default is 20.}

\item{ATRFactor}{Minimum size of candle body compared to the ATR. Default is 1.}

\item{maxuppershadowCL}{Maximum tolerated upper shadow to candle length ratio. Default is 0.1.}

\item{maxlowershadowCL}{Maximum tolerated lower shadow to candle length ratio. Default is 0.1.}
}
\value{
A xts object containing the columns:
\itemize{
\item WhiteMarubozu: TRUE if pattern detected
\item BlackMarubozu: TRUE if pattern detected
}
}
\description{
Identifies Marubozu candlestick patterns in an OHLC price series.
Marubozu candles have long bodies with very small or no shadows.
}
\details{
Number of candle lines: \bold{1}

\strong{White Marubozu:}
A long white candle that has no shadows or only small shadows on either end.

\strong{Black Marubozu:}
A long black candle that has no shadows or only small shadows on either end.

In default settings, the candle \emph{body} length must be greater than the average
true range of last \code{n} periods. The threshold can be varied by \code{ATRFactor}.
}
\examples{
\dontrun{
getSymbols("YHOO", adjust = TRUE)

CSPMarubozu(YHOO)

# Include not-so-long-marubozus
CSPMarubozu(YHOO, ATRFactor = 0.8)

# Filter for white closing marubozus (Cl(TS) = Hi(TS))
CSPMarubozu(YHOO, maxuppershadowCL = 0)[, "WhiteMarubozu"]
}

}
\references{
The following sites were used to code/document this indicator:
\itemize{
\item \url{http://www.candlesticker.com/Bullish.asp}
\item \url{http://www.candlesticker.com/Bearish.asp}
}
}
\seealso{
\code{\link{CandleBodyLength}}
\code{\link[TTR]{ATR}}
}
\author{
Andreas Voellenklee
}
