% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSPNBlended.R
\name{CSPNBlended}
\alias{CSPNBlended}
\title{Calculate Open/High/Low/Close of N Blended Bars}
\usage{
CSPNBlended(TS, N)
}
\arguments{
\item{TS}{xts Time Series containing OHLC prices}

\item{N}{Number of bars to combine into one bar. Must be an integer >= 1.}
}
\value{
A xts object containing the columns:
\itemize{
\item {N}.Blended.Open: Opening price of the \code{N}-th elapsed candle
\item {N}.Blended.High: Highest high of the past \code{N} candles
\item {N}.Blended.Low: Lowest low of the past \code{N} candles
\item {N}.Blended.Close: Close price of the current candle
}
}
\description{
This function calculates OHLC prices of a combined candle representing N bars.
It creates a "blended" candle that aggregates price action over multiple periods.
}
\note{
This function is used by \code{\link{addPriceInfo}} to add price information
of detected candlestick patterns.
}
\examples{
\dontrun{
getSymbols("YHOO", adjust = TRUE)
CSPNBlended(YHOO, N = 3) # combine 3 candles into one
}

}
\seealso{
\code{\link{addPriceInfo}}
}
\author{
Andreas Voellenklee
}
