% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSPNHigherClose.R
\name{CSPNHigherClose}
\alias{CSPNHigherClose}
\title{N Consecutive Higher Close Prices}
\usage{
CSPNHigherClose(TS, N)
}
\arguments{
\item{TS}{xts Time Series containing Close prices}

\item{N}{Number of consecutive candles to test for higher close series}
}
\value{
A xts object containing the column:
\itemize{
\item {N}HigherClose: TRUE if current close is the \code{N}th higher close in a row
}
}
\description{
Identifies sequences of N consecutive higher close prices in a time series.
}
\details{
Only the close price is evaluated. The colors of the candles are ignored.

Identifies when the current close price is higher than the previous close price,
and this pattern continues for N consecutive periods.
}
\examples{
\dontrun{
getSymbols("AAPL", adjust = TRUE)
CSPNHigherClose(AAPL, N = 3) # filter for 3 consecutive higher close
}

}
\author{
Andreas Voellenklee
}
