% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSPNLongBlackCandleBodies.R
\name{CSPNLongBlackCandleBodies}
\alias{CSPNLongBlackCandleBodies}
\title{N Consecutive Long Black Candle Bodies}
\usage{
CSPNLongBlackCandleBodies(TS, N = 2, n = 20, threshold = 1)
}
\arguments{
\item{TS}{xts Time Series containing Open and Close prices}

\item{N}{Number of consecutive candles to test. Default is 2.}

\item{n}{Number of preceding candles to calculate median candle body length. Default is 20.}

\item{threshold}{Minimum candle body length in relation to the median candle body length
of \code{n} preceding candles. Default is 1.}
}
\value{
A xts object containing the column:
\itemize{
\item {N}LongBlackCandleBodies: TRUE if current candle is the \code{N}-th consecutive long black candle body
}
}
\description{
Identifies sequences of N consecutive long black candle bodies in an OHLC price series.
}
\details{
This function identifies when the current candle is the N-th consecutive long black candle body.
A candle body is considered "long" if its length is greater than or equal to the threshold
times the median candle body length of the preceding n candles.
}
\examples{
\dontrun{
getSymbols("YHOO", adjust = TRUE)
CSPNLongBlackCandleBodies(YHOO, N = 4, n = 50, threshold = 1.2)
}

}
\seealso{
\code{\link{CSPLongCandleBody}}
}
\author{
Andreas Voellenklee
}
