% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSPNLongBlackCandles.R
\name{CSPNLongBlackCandles}
\alias{CSPNLongBlackCandles}
\title{N Consecutive Long Black Candles}
\usage{
CSPNLongBlackCandles(TS, N = 2, n = 20, threshold = 1)
}
\arguments{
\item{TS}{xts Time Series containing OHLC prices}

\item{N}{Number of consecutive candles to test. Default is 2.}

\item{n}{Number of preceding candles to calculate median candle length. Default is 20.}

\item{threshold}{Minimum candle length in relation to the median candle length
of \code{n} preceding candles. Default is 1.}
}
\value{
A xts object containing the column:
\itemize{
\item {N}LongBlackCandles: TRUE if current candle is the \code{N}-th consecutive long black candle
}
}
\description{
Identifies sequences of N consecutive long black candles in an OHLC price series.
}
\details{
This function identifies when the current candle is the N-th consecutive long black candle.
A candle is considered "long" if its length is greater than or equal to the threshold
times the median candle length of the preceding n candles.
}
\examples{
\dontrun{
getSymbols("YHOO", adjust = TRUE)
CSPNLongBlackCandles(YHOO, N = 3)
}

}
\seealso{
\code{\link{CSPLongCandle}}
}
\author{
Andreas Voellenklee
}
