% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSPNLowerClose.R
\name{CSPNLowerClose}
\alias{CSPNLowerClose}
\title{N Consecutive Lower Close Prices}
\usage{
CSPNLowerClose(TS, N)
}
\arguments{
\item{TS}{xts Time Series containing Close prices}

\item{N}{Number of consecutive candles to test for lower close series}
}
\value{
A xts object containing the column:
\itemize{
\item {N}LowerClose: TRUE if current close is the \code{N}th lower close in a row
}
}
\description{
Identifies sequences of N consecutive lower close prices in a time series.
}
\details{
Only the close price is evaluated. The colors of the candles are ignored.

Identifies when the current close price is lower than the previous close price,
and this pattern continues for N consecutive periods.
}
\examples{
\dontrun{
getSymbols("YHOO", adjust = TRUE)
CSPNLowerClose(YHOO, N = 4) # filter for 4 consecutive lower close
}

}
\author{
Andreas Voellenklee
}
