% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSPShortCandle.R
\name{CSPShortCandle}
\alias{CSPShortCandle}
\title{Short Candlestick Pattern}
\usage{
CSPShortCandle(TS, n = 20, threshold = 1)
}
\arguments{
\item{TS}{xts Time Series containing OHLC prices}

\item{n}{Number of preceding candles to calculate median candle length. Default is 20.}

\item{threshold}{Maximum candle length in relation to the median candle length
of \code{n} preceding candles. Default is 1.}
}
\value{
A xts object containing the columns:
\itemize{
\item ShortWhiteCandle: TRUE if Short White Candle detected
\item ShortBlackCandle: TRUE if Short Black Candle detected
}
}
\description{
Identifies short candlesticks in an OHLC price series based on their relative length
compared to the median of preceding candles.
}
\details{
Classifies candles as short based on their relative length compared to the median
of the relative candle lengths of the preceding \code{n} candles.

Relative candle length is calculated as \code{(Hi-Lo)/((Hi+Lo)/2)}.
}
\examples{
\dontrun{
getSymbols("YHOO", adjust = TRUE)
CSPShortCandle(YHOO, threshold = 0.5) # filter for very small candles
}

}
\seealso{
\code{\link{CandleLength}}, \code{\link{CSPLongCandle}}
}
\author{
Andreas Voellenklee
}
