% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSPShortCandleBody.R
\name{CSPShortCandleBody}
\alias{CSPShortCandleBody}
\title{Short Candle Body Pattern}
\usage{
CSPShortCandleBody(TS, n = 20, threshold = 1)
}
\arguments{
\item{TS}{xts Time Series containing Open and Close prices}

\item{n}{Number of preceding candles to calculate median candle body length. Default is 20.}

\item{threshold}{Maximum candle body length in relation to the median candle body length
of \code{n} preceding candles. Default is 1.}
}
\value{
A xts object containing the columns:
\itemize{
\item ShortWhiteCandleBody: TRUE if Short White Candle Body detected
\item ShortBlackCandleBody: TRUE if Short Black Candle Body detected
}
}
\description{
Identifies short candle bodies in an OHLC price series based on their relative length
compared to the median of preceding candle bodies.
}
\details{
Classifies candle bodies as short based on their relative length compared to the median
of the relative candle body lengths of the preceding \code{n} candles.
}
\examples{
\dontrun{
getSymbols("YHOO", adjust = TRUE)
CSPShortCandleBody(YHOO)
}

}
\seealso{
\code{\link{CandleBodyLength}}, \code{\link{CSPLongCandle}}
}
\author{
Andreas Voellenklee
}
