% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DonchianChannel.R
\name{DonchianChannel2}
\alias{DonchianChannel2}
\title{Donchian Channel}
\usage{
DonchianChannel2(HL, n = 10)
}
\arguments{
\item{HL}{Object that is coercible to xts or matrix and contains High and Low prices or Close prices.}

\item{n}{Number of periods to include in price band calculation.}
}
\value{
An object of the same class as \code{HL} or a matrix containing the columns:
\itemize{
\item high: The highest high series (lagged by 1 period)
\item mid: The average of \code{high} and \code{low} (lagged by 1 period)
\item low: The lowest low series (lagged by 1 period)
}
}
\description{
Calculates Donchian Channels with a 1-period lag correction, as used in the candlesticks package.
Donchian Channels were created by Richard Donchian and were used to generate buy and
sell signals for the Turtle Trading system.
}
\details{
Donchian Channels consist of two (sometimes three) lines:
\itemize{
\item The top line is the highest high of the past \code{n} periods.
\item The bottom line is the lowest low of the past \code{n} periods.
\item The middle line is the average of the top and bottom lines.
}
}
\note{
The DonchianChannel function in the \code{TTR} package includes the current price bar
for the channel calculation. In contrast, this DonchianChannel2 function calculates
the price bands on the \emph{previous} \code{n} price bars and includes a 1-period lag.
}
\examples{
\dontrun{
getSymbols("YHOO", adjust = TRUE)
dc <- DonchianChannel2(YHOO)
}

}
\references{
The following site(s) were used to code/document this indicator:
\url{https://www.investopedia.com/terms/d/donchianchannels.asp}
}
\seealso{
See \code{\link[TTR]{DonchianChannel}} in the TTR package.
}
\author{
G See
}
\keyword{ts}
