% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HL_Average.R
\name{HL_Average}
\alias{HL_Average}
\title{Calculate HL Average}
\usage{
HL_Average(TS)
}
\arguments{
\item{TS}{An xts time series containing HL prices}
}
\value{
An xts object containing a single column:
\itemize{
\item \code{HL_Average}: The average of High and Low prices
}
}
\description{
Calculates the average value of High and Low prices.
}
\details{
The HL Average is calculated as: (High + Low) / 2
}
\examples{
\dontrun{
getSymbols("YHOO", adjust = TRUE)
head(HL_Average(YHOO))
}

}
\author{
Andreas Voellenklee
}
