% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.HL.R
\name{is.HL}
\alias{is.HL}
\title{Check for High/Low Data}
\usage{
is.HL(x)
}
\arguments{
\item{x}{A data object (typically an xts time series)}
}
\value{
\itemize{
\item \code{is.HL}: TRUE if x includes columns for High and Low
}
}
\description{
Function to check for appropriate High/Low column names within a data object.
}
\details{
The candlestick package uses functions to check for OHLC data from the \pkg{quantmod} package.
The function \code{is.HL} is not included in quantmod, so it is provided here.
}
\examples{
\dontrun{
getSymbols("YHOO")
is.HL(YHOO)
}

}
\seealso{
\code{\link[quantmod]{has.OHLC}}
}
