% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.OC.R
\name{is.OC}
\alias{is.OC}
\title{Check for Open/Close Data}
\usage{
is.OC(x)
}
\arguments{
\item{x}{A data object (typically an xts time series)}
}
\value{
\itemize{
\item \code{is.OC}: TRUE if x includes columns for Open and Close
\item \code{is.HL}: TRUE if x includes columns for High and Low
}
}
\description{
Functions to check for appropriate Open/Close and High/Low column names within a data object.
}
\details{
The candlestick package uses functions to check for OHLC data from the \pkg{quantmod} package.
The functions \code{is.OC} and \code{is.HL} are not included in quantmod, so they are provided here.
}
\examples{
\dontrun{
getSymbols("YHOO")
is.OC(YHOO)
is.HL(YHOO)
}

}
\seealso{
\code{\link[quantmod]{has.OHLC}}
}
