% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LagOHLC.R
\name{LagOHLC}
\alias{LagOHLC}
\title{Lag OHLC Time Series}
\usage{
LagOHLC(TS, k = 1)
}
\arguments{
\item{TS}{An xts time series containing Open, High, Low and Close prices}

\item{k}{Number of periods to lag (default = 1)}
}
\value{
An xts object containing lagged OHLC columns. The original column names are
appended with \code{'.Lag.k'}, where \code{k} is the number of periods lagged.
The returned series maintains the number of observations of the original.
}
\description{
Create a lagged series from OHLC price data.
}
\details{
This function shifts the series k-periods down, prepending NAs to the front of the series.
}
\note{
This function uses the Lag function from the \pkg{stats} package.
}
\examples{
\dontrun{
getSymbols("YHOO")
LagOHLC(YHOO) # Lag OHLC series by one period
LagOHLC(YHOO, k = 1:3) # Lag OHLC series by one, two and three periods
# this will return a 12 columns xts object
}

}
\seealso{
\code{\link[stats]{lag}}, \code{\link{LagOC}}
}
