% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextCandlePosition.R
\name{nextCandlePosition}
\alias{nextCandlePosition}
\title{Position and color of the next candle in a OC price series}
\usage{
nextCandlePosition(TS)
}
\arguments{
\item{TS}{xts Time Series containing Open and Close Prices}
}
\value{
A xts object containing the columns:
\itemize{
\item HigherOpen: TRUE if following candle opens higher than current close
\item LowerOpen: TRUE if following candle opens lower than current close
\item HigherClose: TRUE if following candle closes higher than current close
\item LowerClose: TRUE if following candle closes lower than current close
\item White: TRUE if following candle is white (Cl > Op)
\item Black: TRUE if following candle is black (Cl < Op)
}
}
\description{
Examines open and close price of the following candle relative to the close price
of the current candle. Also returns the color of the following candle.
}
\details{
Sometimes it is suggested to wait for a confirmation in the form of a higher/lower
following close and/or a white/black candle before considering trading the pattern.
This function compares the price levels and returns a xts object with columns of
boolean values for each condition.
}
\note{
This function uses the Next function from the \pkg{quantmod} package.
}
\examples{
\dontrun{
# Filter for inverted hammers in uptrends that are followed by a lower close the next day
getSymbols("YHOO", adjust = TRUE)
CSPInvertedHammer(YHOO) &
  TrendDetectionChannel(YHOO)[, "UpTrend"] &
  nextCandlePosition(YHOO)[, "LowerClose"]
}

}
\seealso{
\code{\link[quantmod]{Next}}
}
\author{
Andreas Voellenklee
}
