% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrendDetectionChannel.R
\name{TrendDetectionChannel}
\alias{TrendDetectionChannel}
\title{Current Trend Detection using Donchian Channel}
\usage{
TrendDetectionChannel(TS, n = 20, DCSector = 1/3)
}
\arguments{
\item{TS}{xts Time Series containing OHLC prices}

\item{n}{Number of periods to calculate the high and low band of the Donchian Channel. Default is 20.}

\item{DCSector}{Sector of Donchian Channel to determine uptrend/downtrend. Default is 1/3.}
}
\value{
A xts object containing the columns:
\itemize{
\item UpTrend: TRUE if uptrend detected
\item NoTrend: TRUE if sideward trend detected
\item DownTrend: TRUE if downtrend detected
\item Trend: +1 for uptrend, 0 for sideward trend, -1 for downtrend
}
}
\description{
Uses the Donchian Channel of \code{n} periods to determine whether a price series
is in uptrend, sideward trend or downtrend.
}
\details{
This function assumes that a price series is in uptrend when a period's price closes
above the upper third of the Donchian Channel of \code{n} periods. If the price closes
below the lower third of the channel, a downtrend is detected. If the price closes
within the middle third of the channel, a sideward trend is detected.

The parameter DCSector can be used to widen/narrow the threshold of up/downtrend detection.
For example, a value of 0.25 anticipates the price for an uptrend to be above the highest
quarter of the Donchian channel.
}
\examples{
\dontrun{
getSymbols("YHOO", adjust = TRUE)

# Create chart of YAHOO
chartSeries(YHOO, subset = "last 1 year", TA = NULL)

# Visualize the result of trend detection in a indicator box
addTA(TrendDetectionChannel(YHOO)[, 4])

# Filter YHOO for Hammer Candlestick Patterns that occur in downtrends
Hammer <- CSPHammer(YHOO) & TrendDetectionChannel(YHOO)[, "DownTrend"]

# How frequent are these hammers?
colSums(Hammer, na.rm = TRUE)
}

}
\seealso{
\code{\link{TrendDetectionSMA}}
}
\author{
Andreas Voellenklee
}
