% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrendDetectionSMA.R
\name{TrendDetectionSMA}
\alias{TrendDetectionSMA}
\title{Current Trend Detection using Simple Moving Average}
\usage{
TrendDetectionSMA(TS, n = 20)
}
\arguments{
\item{TS}{xts Time Series containing OHLC prices}

\item{n}{Number of periods for the SMA to average over. Default is 20.}
}
\value{
A xts object containing the columns:
\itemize{
\item UpTrend: TRUE if uptrend detected (Close > SMA(n))
\item NoTrend: TRUE if sideward trend detected (Close == SMA(n))
\item DownTrend: TRUE if downtrend detected (Close < SMA(n))
\item Trend: +1 for uptrend, 0 for sideward trend, -1 for downtrend
}
}
\description{
Uses the SMA of \code{n} periods to determine whether a price series
is in uptrend or downtrend.
}
\details{
This function assumes that a price series is in uptrend (downtrend) when a period's
price closes above (below) the simple moving average of \code{n} periods.
}
\examples{
\dontrun{
getSymbols("YHOO", adjust = TRUE)

# Create chart of YAHOO
chartSeries(YHOO, subset = "last 1 year", TA = NULL)

# Visualize the result of trend detection in a indicator box
addTA(TrendDetectionSMA(YHOO)[, 4])

# Filter YHOO for Hammer Candlestick Patterns that occur in downtrends
Hammer <- CSPHammer(YHOO) & TrendDetectionSMA(YHOO)[, "DownTrend"]

# How frequent are these hammers?
colSums(Hammer, na.rm = TRUE)
}

}
\seealso{
\code{\link{TrendDetectionChannel}}
}
\author{
Andreas Voellenklee
}
