% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCSPTasukiGap.R
\name{addCSPTasukiGap}
\alias{addCSPTasukiGap}
\title{Upside/Downside Tasuki Gap Candlestick Pattern}
\usage{
addCSPTasukiGap(x, output = c("xts", "tibble", "data.frame"))
}
\arguments{
\item{x}{xts Time Series containing OHLC prices}

\item{output}{Character. Return format: \code{"xts"} (default), \code{"tibble"}, or \code{"data.frame"}.}
}
\value{
A xts object containing the columns:
\itemize{
\item UpsideTasukiGap: TRUE if Upside Tasuki Gap pattern detected
\item DownsideTasukiGap: TRUE if Downside Tasuki Gap pattern detected
}
}
\description{
Identifies Upside and Downside Tasuki Gap patterns in an OHLC price series.
These are three-candle continuation patterns that occur within existing trends.
}
\details{
Number of candle lines: \bold{3}

\strong{Upside Tasuki Gap:}
\itemize{
\item The market is in an uptrend
\item First candle: white candle
\item Second candle: white candle that gaps above the high of the first candle
\item Third candle: black candle that opens within the body of the second candle and closes within the gap between the first and second candle
\item This pattern signals a potential continuation of the uptrend
}

\strong{Downside Tasuki Gap:}
\itemize{
\item The market is in a downtrend
\item First candle: black candle
\item Second candle: black candle that gaps below the low of the first candle
\item Third candle: white candle that opens within the body of the second candle and closes within the gap between the first and second candle
\item This pattern signals a potential continuation of the downtrend
}
}
\note{
The function filters patterns that look like Tasuki gap patterns, without considering
the current trend direction. If only patterns in specific trends should be filtered,
an external trend detection function must be used. See examples.
}
\examples{
\dontrun{
getSymbols("SSRI", adjust = TRUE)
addCSPTasukiGap(SSRI)

# Filter upside Tasuki gaps in uptrends
addCSPTasukiGap(SSRI)[, "UpsideTasukiGap"] &
  TrendDetectionChannel(lag(SSRI, k = 3))[, "UpTrend"]
}

}
\references{
The following sites were used to code/document this indicator:
\itemize{
\item \url{http://www.investopedia.com/terms/u/upside-tasuki-gap.asp#axzz1jtuDsM6i}
\item \url{http://www.investopedia.com/terms/d/downside-tasuki-gap.asp#axzz1jtuDsM6i}
\item \url{http://thepatternsite.com/UpsideTasukiGap.html}
\item \url{http://thepatternsite.com/DownsideTasukiGap.html}
}
}
\seealso{
\code{\link{addCSPGap}}

Other pattern-2bar:
\code{\link[=addCSPDarkCloudCover]{addCSPDarkCloudCover()}},
\code{\link[=addCSPEngulfing]{addCSPEngulfing()}},
\code{\link[=addCSPGap]{addCSPGap()}},
\code{\link[=addCSPHarami]{addCSPHarami()}},
\code{\link[=addCSPInsideDay]{addCSPInsideDay()}},
\code{\link[=addCSPKicking]{addCSPKicking()}},
\code{\link[=addCSPOutsideDay]{addCSPOutsideDay()}},
\code{\link[=addCSPPiercingPattern]{addCSPPiercingPattern()}},
\code{\link[=addCSPStomach]{addCSPStomach()}},
\code{\link[=drawCSPDarkCloudCover]{drawCSPDarkCloudCover()}},
\code{\link[=drawCSPEngulfing]{drawCSPEngulfing()}},
\code{\link[=drawCSPGap]{drawCSPGap()}},
\code{\link[=drawCSPHarami]{drawCSPHarami()}},
\code{\link[=drawCSPInsideDay]{drawCSPInsideDay()}},
\code{\link[=drawCSPKicking]{drawCSPKicking()}},
\code{\link[=drawCSPOutsideDay]{drawCSPOutsideDay()}},
\code{\link[=drawCSPPiercingPattern]{drawCSPPiercingPattern()}},
\code{\link[=drawCSPStomach]{drawCSPStomach()}},
\code{\link[=drawCSPTasukiGap]{drawCSPTasukiGap()}}

Other pattern-bull:
\code{\link[=addCSPDoji]{addCSPDoji()}},
\code{\link[=addCSPEngulfing]{addCSPEngulfing()}},
\code{\link[=addCSPGap]{addCSPGap()}},
\code{\link[=addCSPHammer]{addCSPHammer()}},
\code{\link[=addCSPHarami]{addCSPHarami()}},
\code{\link[=addCSPInvertedHammer]{addCSPInvertedHammer()}},
\code{\link[=addCSPKicking]{addCSPKicking()}},
\code{\link[=addCSPLongCandle]{addCSPLongCandle()}},
\code{\link[=addCSPLongCandleBody]{addCSPLongCandleBody()}},
\code{\link[=addCSPMarubozu]{addCSPMarubozu()}},
\code{\link[=addCSPNHigherClose]{addCSPNHigherClose()}},
\code{\link[=addCSPNLongWhiteCandleBodies]{addCSPNLongWhiteCandleBodies()}},
\code{\link[=addCSPNLongWhiteCandles]{addCSPNLongWhiteCandles()}},
\code{\link[=addCSPPiercingPattern]{addCSPPiercingPattern()}},
\code{\link[=addCSPShortCandle]{addCSPShortCandle()}},
\code{\link[=addCSPShortCandleBody]{addCSPShortCandleBody()}},
\code{\link[=addCSPStar]{addCSPStar()}},
\code{\link[=addCSPStomach]{addCSPStomach()}},
\code{\link[=addCSPThreeInside]{addCSPThreeInside()}},
\code{\link[=addCSPThreeLineStrike]{addCSPThreeLineStrike()}},
\code{\link[=addCSPThreeMethods]{addCSPThreeMethods()}},
\code{\link[=addCSPThreeOutside]{addCSPThreeOutside()}},
\code{\link[=addCSPThreeWhiteSoldiers]{addCSPThreeWhiteSoldiers()}},
\code{\link[=drawCSPDoji]{drawCSPDoji()}},
\code{\link[=drawCSPEngulfing]{drawCSPEngulfing()}},
\code{\link[=drawCSPGap]{drawCSPGap()}},
\code{\link[=drawCSPHammer]{drawCSPHammer()}},
\code{\link[=drawCSPHarami]{drawCSPHarami()}},
\code{\link[=drawCSPInvertedHammer]{drawCSPInvertedHammer()}},
\code{\link[=drawCSPKicking]{drawCSPKicking()}},
\code{\link[=drawCSPLongCandle]{drawCSPLongCandle()}},
\code{\link[=drawCSPLongCandleBody]{drawCSPLongCandleBody()}},
\code{\link[=drawCSPMarubozu]{drawCSPMarubozu()}},
\code{\link[=drawCSPNHigherClose]{drawCSPNHigherClose()}},
\code{\link[=drawCSPNLongWhiteCandleBodies]{drawCSPNLongWhiteCandleBodies()}},
\code{\link[=drawCSPNLongWhiteCandles]{drawCSPNLongWhiteCandles()}},
\code{\link[=drawCSPPiercingPattern]{drawCSPPiercingPattern()}},
\code{\link[=drawCSPShortCandle]{drawCSPShortCandle()}},
\code{\link[=drawCSPShortCandleBody]{drawCSPShortCandleBody()}},
\code{\link[=drawCSPStar]{drawCSPStar()}},
\code{\link[=drawCSPStomach]{drawCSPStomach()}},
\code{\link[=drawCSPTasukiGap]{drawCSPTasukiGap()}},
\code{\link[=drawCSPThreeInside]{drawCSPThreeInside()}},
\code{\link[=drawCSPThreeLineStrike]{drawCSPThreeLineStrike()}},
\code{\link[=drawCSPThreeMethods]{drawCSPThreeMethods()}},
\code{\link[=drawCSPThreeOutside]{drawCSPThreeOutside()}},
\code{\link[=drawCSPThreeWhiteSoldiers]{drawCSPThreeWhiteSoldiers()}}

Other pattern-bear:
\code{\link[=addCSPDarkCloudCover]{addCSPDarkCloudCover()}},
\code{\link[=addCSPDoji]{addCSPDoji()}},
\code{\link[=addCSPEngulfing]{addCSPEngulfing()}},
\code{\link[=addCSPGap]{addCSPGap()}},
\code{\link[=addCSPHangingMan]{addCSPHangingMan()}},
\code{\link[=addCSPHarami]{addCSPHarami()}},
\code{\link[=addCSPKicking]{addCSPKicking()}},
\code{\link[=addCSPLongCandle]{addCSPLongCandle()}},
\code{\link[=addCSPLongCandleBody]{addCSPLongCandleBody()}},
\code{\link[=addCSPMarubozu]{addCSPMarubozu()}},
\code{\link[=addCSPNLongBlackCandleBodies]{addCSPNLongBlackCandleBodies()}},
\code{\link[=addCSPNLongBlackCandles]{addCSPNLongBlackCandles()}},
\code{\link[=addCSPNLowerClose]{addCSPNLowerClose()}},
\code{\link[=addCSPShootingStar]{addCSPShootingStar()}},
\code{\link[=addCSPShortCandle]{addCSPShortCandle()}},
\code{\link[=addCSPShortCandleBody]{addCSPShortCandleBody()}},
\code{\link[=addCSPStar]{addCSPStar()}},
\code{\link[=addCSPStomach]{addCSPStomach()}},
\code{\link[=addCSPThreeBlackCrows]{addCSPThreeBlackCrows()}},
\code{\link[=addCSPThreeInside]{addCSPThreeInside()}},
\code{\link[=addCSPThreeLineStrike]{addCSPThreeLineStrike()}},
\code{\link[=addCSPThreeMethods]{addCSPThreeMethods()}},
\code{\link[=addCSPThreeOutside]{addCSPThreeOutside()}},
\code{\link[=drawCSPDarkCloudCover]{drawCSPDarkCloudCover()}},
\code{\link[=drawCSPDoji]{drawCSPDoji()}},
\code{\link[=drawCSPEngulfing]{drawCSPEngulfing()}},
\code{\link[=drawCSPGap]{drawCSPGap()}},
\code{\link[=drawCSPHangingMan]{drawCSPHangingMan()}},
\code{\link[=drawCSPHarami]{drawCSPHarami()}},
\code{\link[=drawCSPKicking]{drawCSPKicking()}},
\code{\link[=drawCSPLongCandle]{drawCSPLongCandle()}},
\code{\link[=drawCSPLongCandleBody]{drawCSPLongCandleBody()}},
\code{\link[=drawCSPMarubozu]{drawCSPMarubozu()}},
\code{\link[=drawCSPNLongBlackCandleBodies]{drawCSPNLongBlackCandleBodies()}},
\code{\link[=drawCSPNLongBlackCandles]{drawCSPNLongBlackCandles()}},
\code{\link[=drawCSPNLowerClose]{drawCSPNLowerClose()}},
\code{\link[=drawCSPShootingStar]{drawCSPShootingStar()}},
\code{\link[=drawCSPShortCandle]{drawCSPShortCandle()}},
\code{\link[=drawCSPShortCandleBody]{drawCSPShortCandleBody()}},
\code{\link[=drawCSPStar]{drawCSPStar()}},
\code{\link[=drawCSPStomach]{drawCSPStomach()}},
\code{\link[=drawCSPTasukiGap]{drawCSPTasukiGap()}},
\code{\link[=drawCSPThreeBlackCrows]{drawCSPThreeBlackCrows()}},
\code{\link[=drawCSPThreeInside]{drawCSPThreeInside()}},
\code{\link[=drawCSPThreeLineStrike]{drawCSPThreeLineStrike()}},
\code{\link[=drawCSPThreeMethods]{drawCSPThreeMethods()}},
\code{\link[=drawCSPThreeOutside]{drawCSPThreeOutside()}}
}
\author{
Andreas Voellenklee
}
\concept{pattern-2bar}
\concept{pattern-bear}
\concept{pattern-bull}
