% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawCSPPiercingPattern.R
\name{drawCSPPiercingPattern}
\alias{drawCSPPiercingPattern}
\title{Add CSP Piercing Pattern Points to Candlestick Chart}
\usage{
drawCSPPiercingPattern(
  eCandleSticks_result,
  csp_piercing_pattern_result,
  piercing_pattern_color = "blue",
  point_size = 3,
  piercing_pattern_shape = 21,
  point_alpha = 0.8,
  mark_at_close = TRUE
)
}
\arguments{
\item{eCandleSticks_result}{The result object returned by eCandleSticks function}

\item{csp_piercing_pattern_result}{The result object from CSP Piercing Pattern analysis,
which should be a data frame containing a 'Date' column and a logical column 'PiercingPattern'}

\item{piercing_pattern_color}{Color for Piercing Pattern points. Default "blue".}

\item{point_size}{Size for the Piercing Pattern points. Default 3.}

\item{piercing_pattern_shape}{Shape for Piercing Pattern points. Default 21 (filled circle with border).}

\item{point_alpha}{Alpha transparency for the Piercing Pattern points. Default 0.8.}

\item{mark_at_close}{Whether to mark at the close price. Default TRUE.}
}
\value{
A modified eCandleSticks result list with Piercing Pattern points added to the price plot
and the combined plot updated accordingly.
}
\description{
This function adds CSP Piercing Pattern points to a candlestick chart
created by eCandleSticks using the results from a CSP Piercing Pattern analysis,
and recombines it with the volume subplot if it exists.
}
\examples{
\dontrun{
library(quantmod)
getSymbols("AAPL", src = "yahoo", from = "2023-01-01", to = "2025-09-08")

# Create candlestick chart
result <- eCandleSticks(AAPL)

# Get CSP Piercing Pattern results
csp_piercing_pattern_data <- addCSPPiercingPattern(AAPL)

# Add Piercing Pattern points
result_with_piercing_pattern <- drawCSPPiercingPattern(result, csp_piercing_pattern_data)

# Display the combined plot with Piercing Pattern points
print(result_with_piercing_pattern$combined_plot)
}
}
\seealso{
Other pattern-2bar:
\code{\link[=addCSPCounterAttack]{addCSPCounterAttack()}},
\code{\link[=addCSPDarkCloudCover]{addCSPDarkCloudCover()}},
\code{\link[=addCSPDojiStar]{addCSPDojiStar()}},
\code{\link[=addCSPEngulfing]{addCSPEngulfing()}},
\code{\link[=addCSPGap]{addCSPGap()}},
\code{\link[=addCSPHarami]{addCSPHarami()}},
\code{\link[=addCSPHomingPigeon]{addCSPHomingPigeon()}},
\code{\link[=addCSPInNeck]{addCSPInNeck()}},
\code{\link[=addCSPInsideDay]{addCSPInsideDay()}},
\code{\link[=addCSPKicking]{addCSPKicking()}},
\code{\link[=addCSPMatchingLow]{addCSPMatchingLow()}},
\code{\link[=addCSPOnNeck]{addCSPOnNeck()}},
\code{\link[=addCSPOutsideDay]{addCSPOutsideDay()}},
\code{\link[=addCSPPiercingPattern]{addCSPPiercingPattern()}},
\code{\link[=addCSPSeparatingLines]{addCSPSeparatingLines()}},
\code{\link[=addCSPStomach]{addCSPStomach()}},
\code{\link[=addCSPTasukiGap]{addCSPTasukiGap()}},
\code{\link[=addCSPThrusting]{addCSPThrusting()}},
\code{\link[=drawCSPCounterAttack]{drawCSPCounterAttack()}},
\code{\link[=drawCSPDarkCloudCover]{drawCSPDarkCloudCover()}},
\code{\link[=drawCSPDojiStar]{drawCSPDojiStar()}},
\code{\link[=drawCSPEngulfing]{drawCSPEngulfing()}},
\code{\link[=drawCSPGap]{drawCSPGap()}},
\code{\link[=drawCSPHarami]{drawCSPHarami()}},
\code{\link[=drawCSPHomingPigeon]{drawCSPHomingPigeon()}},
\code{\link[=drawCSPInNeck]{drawCSPInNeck()}},
\code{\link[=drawCSPInsideDay]{drawCSPInsideDay()}},
\code{\link[=drawCSPKicking]{drawCSPKicking()}},
\code{\link[=drawCSPMatchingLow]{drawCSPMatchingLow()}},
\code{\link[=drawCSPOnNeck]{drawCSPOnNeck()}},
\code{\link[=drawCSPOutsideDay]{drawCSPOutsideDay()}},
\code{\link[=drawCSPSeparatingLines]{drawCSPSeparatingLines()}},
\code{\link[=drawCSPStomach]{drawCSPStomach()}},
\code{\link[=drawCSPTasukiGap]{drawCSPTasukiGap()}},
\code{\link[=drawCSPThrusting]{drawCSPThrusting()}}

Other pattern-bull:
\code{\link[=addCSPAbandonedBaby]{addCSPAbandonedBaby()}},
\code{\link[=addCSPBeltHold]{addCSPBeltHold()}},
\code{\link[=addCSPBreakaway]{addCSPBreakaway()}},
\code{\link[=addCSPClosingMarubozu]{addCSPClosingMarubozu()}},
\code{\link[=addCSPConcealBabySwallow]{addCSPConcealBabySwallow()}},
\code{\link[=addCSPCounterAttack]{addCSPCounterAttack()}},
\code{\link[=addCSPDoji]{addCSPDoji()}},
\code{\link[=addCSPDojiStar]{addCSPDojiStar()}},
\code{\link[=addCSPEngulfing]{addCSPEngulfing()}},
\code{\link[=addCSPGap]{addCSPGap()}},
\code{\link[=addCSPGapSideSideWhite]{addCSPGapSideSideWhite()}},
\code{\link[=addCSPHammer]{addCSPHammer()}},
\code{\link[=addCSPHarami]{addCSPHarami()}},
\code{\link[=addCSPHomingPigeon]{addCSPHomingPigeon()}},
\code{\link[=addCSPInvertedHammer]{addCSPInvertedHammer()}},
\code{\link[=addCSPKicking]{addCSPKicking()}},
\code{\link[=addCSPLadderBottom]{addCSPLadderBottom()}},
\code{\link[=addCSPLongCandle]{addCSPLongCandle()}},
\code{\link[=addCSPLongCandleBody]{addCSPLongCandleBody()}},
\code{\link[=addCSPMarubozu]{addCSPMarubozu()}},
\code{\link[=addCSPMatchingLow]{addCSPMatchingLow()}},
\code{\link[=addCSPNHigherClose]{addCSPNHigherClose()}},
\code{\link[=addCSPNLongWhiteCandleBodies]{addCSPNLongWhiteCandleBodies()}},
\code{\link[=addCSPNLongWhiteCandles]{addCSPNLongWhiteCandles()}},
\code{\link[=addCSPPiercingPattern]{addCSPPiercingPattern()}},
\code{\link[=addCSPSeparatingLines]{addCSPSeparatingLines()}},
\code{\link[=addCSPShortCandle]{addCSPShortCandle()}},
\code{\link[=addCSPShortCandleBody]{addCSPShortCandleBody()}},
\code{\link[=addCSPSpinningTop]{addCSPSpinningTop()}},
\code{\link[=addCSPStar]{addCSPStar()}},
\code{\link[=addCSPStickSandwich]{addCSPStickSandwich()}},
\code{\link[=addCSPStomach]{addCSPStomach()}},
\code{\link[=addCSPTakuri]{addCSPTakuri()}},
\code{\link[=addCSPTasukiGap]{addCSPTasukiGap()}},
\code{\link[=addCSPThreeInside]{addCSPThreeInside()}},
\code{\link[=addCSPThreeLineStrike]{addCSPThreeLineStrike()}},
\code{\link[=addCSPThreeMethods]{addCSPThreeMethods()}},
\code{\link[=addCSPThreeOutside]{addCSPThreeOutside()}},
\code{\link[=addCSPThreeStarsInSouth]{addCSPThreeStarsInSouth()}},
\code{\link[=addCSPThreeWhiteSoldiers]{addCSPThreeWhiteSoldiers()}},
\code{\link[=addCSPTristar]{addCSPTristar()}},
\code{\link[=addCSPUnique3River]{addCSPUnique3River()}},
\code{\link[=drawCSPAbandonedBaby]{drawCSPAbandonedBaby()}},
\code{\link[=drawCSPBeltHold]{drawCSPBeltHold()}},
\code{\link[=drawCSPBreakaway]{drawCSPBreakaway()}},
\code{\link[=drawCSPClosingMarubozu]{drawCSPClosingMarubozu()}},
\code{\link[=drawCSPConcealBabySwallow]{drawCSPConcealBabySwallow()}},
\code{\link[=drawCSPCounterAttack]{drawCSPCounterAttack()}},
\code{\link[=drawCSPDoji]{drawCSPDoji()}},
\code{\link[=drawCSPDojiStar]{drawCSPDojiStar()}},
\code{\link[=drawCSPEngulfing]{drawCSPEngulfing()}},
\code{\link[=drawCSPGap]{drawCSPGap()}},
\code{\link[=drawCSPGapSideSideWhite]{drawCSPGapSideSideWhite()}},
\code{\link[=drawCSPHammer]{drawCSPHammer()}},
\code{\link[=drawCSPHarami]{drawCSPHarami()}},
\code{\link[=drawCSPHomingPigeon]{drawCSPHomingPigeon()}},
\code{\link[=drawCSPInvertedHammer]{drawCSPInvertedHammer()}},
\code{\link[=drawCSPKicking]{drawCSPKicking()}},
\code{\link[=drawCSPLadderBottom]{drawCSPLadderBottom()}},
\code{\link[=drawCSPLongCandle]{drawCSPLongCandle()}},
\code{\link[=drawCSPLongCandleBody]{drawCSPLongCandleBody()}},
\code{\link[=drawCSPMarubozu]{drawCSPMarubozu()}},
\code{\link[=drawCSPMatchingLow]{drawCSPMatchingLow()}},
\code{\link[=drawCSPNHigherClose]{drawCSPNHigherClose()}},
\code{\link[=drawCSPNLongWhiteCandleBodies]{drawCSPNLongWhiteCandleBodies()}},
\code{\link[=drawCSPNLongWhiteCandles]{drawCSPNLongWhiteCandles()}},
\code{\link[=drawCSPSeparatingLines]{drawCSPSeparatingLines()}},
\code{\link[=drawCSPShortCandle]{drawCSPShortCandle()}},
\code{\link[=drawCSPShortCandleBody]{drawCSPShortCandleBody()}},
\code{\link[=drawCSPSpinningTop]{drawCSPSpinningTop()}},
\code{\link[=drawCSPStar]{drawCSPStar()}},
\code{\link[=drawCSPStickSandwich]{drawCSPStickSandwich()}},
\code{\link[=drawCSPStomach]{drawCSPStomach()}},
\code{\link[=drawCSPTakuri]{drawCSPTakuri()}},
\code{\link[=drawCSPTasukiGap]{drawCSPTasukiGap()}},
\code{\link[=drawCSPThreeInside]{drawCSPThreeInside()}},
\code{\link[=drawCSPThreeLineStrike]{drawCSPThreeLineStrike()}},
\code{\link[=drawCSPThreeMethods]{drawCSPThreeMethods()}},
\code{\link[=drawCSPThreeOutside]{drawCSPThreeOutside()}},
\code{\link[=drawCSPThreeStarsInSouth]{drawCSPThreeStarsInSouth()}},
\code{\link[=drawCSPThreeWhiteSoldiers]{drawCSPThreeWhiteSoldiers()}},
\code{\link[=drawCSPTristar]{drawCSPTristar()}},
\code{\link[=drawCSPUnique3River]{drawCSPUnique3River()}}
}
\concept{pattern-2bar}
\concept{pattern-bull}
